refactor(cli): extract 25 command groups into cli_*.py (3a slice 2)
Decompose the cli.py god-module (9138 LOC) by pulling 25 Click groups into
self-contained per-group modules following the cli_graph.py pattern: each
imports shared state from cli_common and registers onto the root via
add_command. cli.py drops to ~5560 LOC.

Extracted + wired (22): feedback, repo, recall-daemon, embed-daemon, as-of,
session, temporal, consolidate, contextual, links, version, query, backup,
sync, encrypt, share, analytics, import, export, multimodal, collaborative,
contradict.

Extracted but intentionally NOT registered (3): federation, lifecycle,
suggest — these are "brain-like" internal surfaces that
test_brain_like_cli_groups_are_not_public asserts must stay out of
cli.commands (Synapse owns orchestration). They live in their modules,
unwired, preserving prior behavior.

Helper relocation to avoid cross-module coupling:
- _short + _parse_as_of_date -> cli_common (shared by an extracted group and
  a top-level command that stays in cli.py).
- _repo_index_operational_receipt -> cli_repo, with its deps decoupled
  (_now_iso -> util.utc_now_iso, _backend_native_trace_id -> inline
  os.environ read, _stable_json_hash -> util.stable_hash).
- _portable_backup -> cli_backup.

Deferred: profile + backend-native groups stay in cli.py (entangled with the
_profile_status_report/_profile_repair_plan diagnostic cluster shared with
doctor/stats) — to be handled with the diag helpers in a later slice.

Verification: ruff clean, mypy clean (83 files), pytest 736 passed,
--help smoke on every extracted subgroup.
