+++ src/memo/cli.py
@@ -42,15 +42,36 @@ import click
 from rich.panel import Panel
 from rich.table import Table
 
-from memo.cli_common import console
+from memo.cli_analytics import analytics_group
+from memo.cli_as_of import as_of_group
+from memo.cli_backup import backup_group
+from memo.cli_collaborative import collaborative_group
+from memo.cli_common import _parse_as_of_date, _short, console
 from memo.cli_common import get_memory as _get_memory
+from memo.cli_consolidate import consolidate_group
+from memo.cli_contextual import contextual_group
+from memo.cli_contradict import contradict_group
+from memo.cli_embed_daemon import embed_daemon_group
+from memo.cli_encrypt import encrypt_group
+from memo.cli_export import export_group
+from memo.cli_feedback import feedback_group
 from memo.cli_graph import graph_group
+from memo.cli_import import import_group
+from memo.cli_links import links_group
+from memo.cli_multimodal import multimodal_group
+from memo.cli_query import query_group
+from memo.cli_recall_daemon import recall_daemon_group
+from memo.cli_repo import repo_group
+from memo.cli_session import session_group
+from memo.cli_share import share_group
+from memo.cli_sync import sync_group
+from memo.cli_temporal import temporal_group
+from memo.cli_version import version_group
 from memo.config import MODEL_PROFILES, Config
 
 # Imported at module scope (not lazily) so tests can `patch("memo.cli.run_picker", ...)`.
 # `run_picker` itself defers the heavy `questionary` import until called.
 from memo.setup import run_picker, write_config_file
-from memo.util import stable_hash as _stable_json_hash
 
 
 def _now_iso() -> str:
@@ -59,46 +80,6 @@ def _now_iso() -> str:
 
 
 
-def _repo_index_operational_receipt(out: dict[str, Any]) -> dict[str, Any]:
-    name = str(out.get("name") or "repo")
-    commit = str(out.get("commit_sha") or "")
-    safe_name = re.sub(r"[^A-Za-z0-9_.-]+", "-", name).strip("-") or "repo"
-    receipt_id = f"{safe_name}/{commit[:12] or 'unknown'}"
-    status = "partial" if int(out.get("errors") or 0) else "ok"
-    receipt = {
-        "schema": "memo.operational_receipt.v1",
-        "source": "memo",
-        "operation": "repo_index",
-        "status": status,
-        "uri": f"memo://repo-index/{receipt_id}",
-        "generated_at": _now_iso(),
-        "trace_id": _backend_native_trace_id(),
-        "repo": {
-            "id": out.get("repo_id") or "",
-            "name": name,
-            "url": out.get("url") or "",
-            "ref": out.get("ref") or "",
-            "commit_sha": commit,
-            "semantic_status": out.get("semantic_status") or "",
-        },
-        "counts": {
-            "checked_files": int(out.get("checked_files") or 0),
-            "indexed_files": int(out.get("indexed_files") or 0),
-            "unchanged_files": int(out.get("unchanged_files") or 0),
-            "deleted_files": int(out.get("deleted_files") or 0),
-            "indexed_chunks": int(out.get("indexed_chunks") or 0),
-            "indexed_lines": int(out.get("indexed_lines") or 0),
-            "embedded_chunks": int(out.get("embedded_chunks") or 0),
-            "pending_chunks": int(out.get("pending_chunks") or 0),
-            "errors": int(out.get("errors") or 0),
-        },
-        "provenance": {
-            "memo_uri": f"memo://repo/{out.get('repo_id') or safe_name}",
-            "clone_path": out.get("clone_path") or "",
-        },
-    }
-    receipt["content_hash"] = _stable_json_hash(receipt)
-    return receipt
 
 
 def _managed_sqlite_dbs(cfg: Config) -> list[tuple[str, Path]]:
@@ -616,6 +597,28 @@ def cli(ctx: click.Context) -> None:
 
 # Command groups extracted from this module live in cli_*.py and register here.
 cli.add_command(graph_group)
+cli.add_command(feedback_group)
+cli.add_command(repo_group)
+cli.add_command(recall_daemon_group)
+cli.add_command(embed_daemon_group)
+cli.add_command(as_of_group)
+cli.add_command(session_group)
+cli.add_command(temporal_group)
+cli.add_command(consolidate_group)
+cli.add_command(contextual_group)
+cli.add_command(links_group)
+cli.add_command(version_group)
+cli.add_command(query_group)
+cli.add_command(backup_group)
+cli.add_command(sync_group)
+cli.add_command(encrypt_group)
+cli.add_command(share_group)
+cli.add_command(analytics_group)
+cli.add_command(import_group)
+cli.add_command(export_group)
+cli.add_command(multimodal_group)
+cli.add_command(collaborative_group)
+cli.add_command(contradict_group)
 
 
 # Subcommands that must NEVER trigger the first-run picker — either
@@ -1863,409 +1866,8 @@ def reindex(force: bool, as_json: bool) -> None:
     )
 
 
-@cli.group(name="feedback")
-def feedback_group() -> None:
-    """Per-source 👍/👎 feedback used to teach the retriever which
-    memorias to surface (or hide) for queries similar to the one being
-    voted on."""
 
 
-@feedback_group.command(name="record")
-@click.argument("source_id")
-@click.option("--query", "query_text", required=True, help="Query text the feedback applies to.")
-@click.option("--rating", required=True, type=click.Choice(["up", "down"]), help="up = boost, down = exclude for similar queries.")
-@click.option("--as-json", is_flag=True)
-def feedback_record_cmd(source_id: str, query_text: str, rating: str, as_json: bool) -> None:
-    """Record a 👍/👎 vote on SOURCE_ID for QUERY. Embeds QUERY so future
-    semantically-similar queries inherit the vote.
-
-    SOURCE_ID may be a full meta.id or a unique prefix (>= 4 chars)."""
-    from memo.memory import Memory
-    cfg = Config.from_env()
-    mem = Memory(cfg)
-    rid = mem.feedback_record(source_id, query_text=query_text, rating=rating)
-    if as_json:
-        click.echo(json.dumps(rid, ensure_ascii=False))
-        return
-    console.print(
-        f"[green]ok[/green] feedback_id={rid['feedback_id']} "
-        f"source={rid['source_id'][:8]} rating={rid['rating']}"
-    )
-
-
-@feedback_group.command(name="list")
-@click.option("--source", "source_id", default=None, help="Filter by source id (prefix ok).")
-@click.option("--limit", default=50, type=int)
-@click.option("--as-json", is_flag=True)
-def feedback_list_cmd(source_id: str | None, limit: int, as_json: bool) -> None:
-    """List recorded feedback rows, newest first."""
-    from memo.memory import Memory
-    cfg = Config.from_env()
-    mem = Memory(cfg)
-    rows = mem.feedback_list(source_id=source_id, limit=limit)
-    if as_json:
-        click.echo(json.dumps(rows, ensure_ascii=False, indent=2))
-        return
-    if not rows:
-        console.print("[dim]no feedback recorded[/dim]")
-        return
-    for r in rows:
-        rating_disp = "👍" if int(r["rating"]) > 0 else "👎"
-        console.print(
-            f"{rating_disp} [cyan]{r['source_id'][:8]}[/cyan] "
-            f"[dim]{r['created_at']}[/dim]  q={r['query_text']!r}"
-        )
-
-
-@feedback_group.command(name="clear")
-@click.argument("source_id")
-@click.option("--yes", is_flag=True, help="Skip confirmation prompt.")
-def feedback_clear_cmd(source_id: str, yes: bool) -> None:
-    """Drop ALL feedback rows for SOURCE_ID. Cannot be undone."""
-    if not yes and not click.confirm(f"Drop all feedback for {source_id}?"):
-        return
-    from memo.memory import Memory
-    cfg = Config.from_env()
-    mem = Memory(cfg)
-    n = mem.feedback_clear(source_id)
-    console.print(f"[green]ok[/green] deleted {n} row(s)")
-
-
-@cli.group(name="repo")
-def repo_group() -> None:
-    """Index and search Git repositories."""
-
-
-def _run_with_repo_progress(
-    thunk: Callable[[Callable[[str, dict[str, Any]], None]], dict[str, Any]],
-) -> dict[str, Any]:
-    from rich.progress import (
-        BarColumn,
-        Progress,
-        SpinnerColumn,
-        TextColumn,
-        TimeElapsedColumn,
-    )
-
-    progress_state: dict[str, Any] = {
-        "files_task": None,
-        "semantic_task": None,
-    }
-
-    with Progress(
-        SpinnerColumn(),
-        TextColumn("[progress.description]{task.description}"),
-        BarColumn(),
-        TextColumn("{task.completed}/{task.total}"),
-        TimeElapsedColumn(),
-        console=console,
-    ) as progress:
-        def _on_progress(event: str, data: dict[str, Any]) -> None:
-            if event == "clone_start":
-                console.print(f"[dim]clone/fetch[/dim] {data['url']}")
-            elif event == "clone_done":
-                console.print(f"[dim]commit[/dim] {str(data['commit_sha'])[:8]}")
-            elif event == "scan_start":
-                progress_state["files_task"] = progress.add_task(
-                    "scan/index files",
-                    total=int(data.get("total") or 0),
-                )
-            elif event in {"file_skipped", "file_indexed", "file_error"}:
-                task_id = progress_state.get("files_task")
-                if task_id is not None:
-                    progress.advance(task_id)
-            elif event == "write_start":
-                console.print(
-                    "[dim]write sqlite[/dim] "
-                    f"flush_batch={data.get('flush_batch')}"
-                )
-            elif event == "write_done":
-                console.print(
-                    "[dim]write sqlite done[/dim] "
-                    f"files={data.get('files')} chunks={data.get('chunks')} "
-                    f"lines={data.get('lines')}"
-                )
-            elif event == "semantic_prepare":
-                console.print(
-                    "[dim]embedder[/dim] preparing pending chunks "
-                    f"for {data.get('repo') or 'repo'}"
-                )
-            elif event == "semantic_start":
-                total = int(data.get("chunks") or 0)
-                if total:
-                    repo = data.get("repo") or "repo"
-                    console.print(
-                        "[dim]embedder[/dim] "
-                        f"{total} pending chunks; use --no-embeddings to skip"
-                    )
-                    progress_state["semantic_task"] = progress.add_task(
-                        f"embedder {repo}",
-                        total=total,
-                    )
-                else:
-                    console.print("[dim]embedder[/dim] no pending chunks")
-            elif event == "semantic_batch":
-                task_id = progress_state.get("semantic_task")
-                if task_id is not None:
-                    progress.update(task_id, completed=int(data.get("completed") or 0))
-            elif event == "semantic_done":
-                task_id = progress_state.get("semantic_task")
-                if task_id is not None:
-                    total = int(data.get("total") or data.get("embedded") or 0)
-                    progress.update(task_id, completed=total)
-
-        return thunk(_on_progress)
-
-
-@repo_group.command(name="index")
-@click.argument("url")
-@click.option("--name", default=None, help="Stable repo label. Defaults to repo basename.")
-@click.option("--ref", "ref_", default=None, help="Branch, tag, or commit. Default: current/default HEAD.")
-@click.option("--force", is_flag=True, help="Re-index even when commit/file hashes are unchanged.")
-@click.option("--no-embeddings", is_flag=True, help="Write exact line/BM25 index only; run `memo repo embed` later.")
-@click.option("--include", multiple=True, help="Glob to include. Repeatable. Default: all text files.")
-@click.option("--exclude", multiple=True, help="Glob to exclude. Repeatable.")
-@click.option("--max-file-bytes", default=None, type=int, help="Skip files above this byte size.")
-@click.option("--json", "as_json", is_flag=True)
-def repo_index_cmd(
-    url: str,
-    name: str | None,
-    ref_: str | None,
-    force: bool,
-    no_embeddings: bool,
-    include: tuple[str, ...],
-    exclude: tuple[str, ...],
-    max_file_bytes: int | None,
-    as_json: bool,
-) -> None:
-    """Clone/fetch URL and index included text files line-by-line."""
-    from memo.memory import Memory
-
-    mem = Memory(Config.from_env())
-    try:
-        if as_json:
-            out = mem.repo_index(
-                url,
-                name=name,
-                ref=ref_,
-                force=force,
-                with_embeddings=not no_embeddings,
-                include=list(include),
-                exclude=list(exclude),
-                max_file_bytes=max_file_bytes,
-            )
-        else:
-            def _run(progress: Callable[[str, dict[str, Any]], None]) -> dict[str, Any]:
-                out = mem.repo_index(
-                    url,
-                    name=name,
-                    ref=ref_,
-                    force=force,
-                    with_embeddings=not no_embeddings,
-                    include=list(include),
-                    exclude=list(exclude),
-                    max_file_bytes=max_file_bytes,
-                    progress=progress,
-                )
-                return out
-
-            out = _run_with_repo_progress(_run)
-    except Exception as exc:
-        raise click.ClickException(str(exc)) from exc
-
-    receipt = _repo_index_operational_receipt(out)
-    out = {**out, "operational_receipt": receipt}
-    if as_json:
-        click.echo(json.dumps(out, ensure_ascii=False, indent=2))
-        return
-    console.print(
-        f"[green]done[/green] repo={out['name']} commit={out['commit_sha'][:8]} "
-        f"files={out['indexed_files']} unchanged={out['unchanged_files']} "
-        f"deleted={out['deleted_files']} chunks={out['indexed_chunks']} "
-        f"lines={out['indexed_lines']} embedded={out['embedded_chunks']} "
-        f"model={out.get('model_chunks', 0)} cached={out.get('cached_chunks', 0)} "
-        f"pending={out['pending_chunks']} status={out['semantic_status']} "
-        f"errors={out['errors']}"
-    )
-    console.print(f"[dim]receipt[/dim] {receipt['uri']}")
-
-
-@repo_group.command(name="embed")
-@click.argument("repo")
-@click.option("--force", is_flag=True, help="Re-embed all chunks, not only pending chunks.")
-@click.option("--json", "as_json", is_flag=True)
-def repo_embed_cmd(repo: str, force: bool, as_json: bool) -> None:
-    """Embed pending chunks for an already indexed repo."""
-    from memo.memory import Memory
-
-    mem = Memory(Config.from_env())
-    try:
-        if as_json:
-            out = mem.repo_embed(repo, force=force)
-        else:
-            out = _run_with_repo_progress(
-                lambda progress: mem.repo_embed(repo, force=force, progress=progress),
-            )
-    except Exception as exc:
-        raise click.ClickException(str(exc)) from exc
-
-    if as_json:
-        click.echo(json.dumps(out, ensure_ascii=False, indent=2))
-        return
-    console.print(
-        f"[green]done[/green] repo={out['name']} embedded={out['embedded_chunks']} "
-        f"model={out.get('model_chunks', 0)} cached={out.get('cached_chunks', 0)} "
-        f"total={out['total_chunks']} pending={out['pending_chunks']} "
-        f"status={out['semantic_status']}"
-    )
-
-
-@repo_group.command(name="status")
-@click.argument("repo")
-@click.option("--json", "as_json", is_flag=True)
-def repo_status_cmd(repo: str, as_json: bool) -> None:
-    """Show exact and semantic index status for one repo."""
-    from memo.memory import Memory
-
-    mem = Memory(Config.from_env())
-    out = mem.repo_status(repo)
-    if out is None:
-        console.print(f"[red]not found:[/red] {repo}")
-        sys.exit(1)
-    if as_json:
-        click.echo(json.dumps(out, ensure_ascii=False, indent=2))
-        return
-    console.print(
-        f"[bold]{out['name']}[/bold] {out['commit_sha'][:8]} "
-        f"status={out['semantic_status']} files={out['files']} "
-        f"lines={out['lines']} chunks={out['chunks']} "
-        f"embedded={out['embedded_chunks']} pending={out['pending_chunks']}"
-    )
-
-
-@repo_group.command(name="list")
-@click.option("--limit", default=100, type=int, show_default=True)
-@click.option("--json", "as_json", is_flag=True)
-def repo_list_cmd(limit: int, as_json: bool) -> None:
-    """List indexed repositories."""
-    from memo.memory import Memory
-
-    mem = Memory(Config.from_env())
-    rows = mem.repo_list(limit=limit)
-    if as_json:
-        click.echo(json.dumps(rows, ensure_ascii=False, indent=2))
-        return
-    if not rows:
-        console.print("[dim]no repos indexed[/dim]")
-        return
-    tbl = Table(show_lines=False, expand=True)
-    tbl.add_column("indexed", width=20)
-    tbl.add_column("name", width=18)
-    tbl.add_column("ref", width=12)
-    tbl.add_column("commit", width=10)
-    tbl.add_column("url", overflow="fold")
-    for r in rows:
-        tbl.add_row(
-            str(r.get("indexed_at", ""))[:19],
-            r["name"],
-            r["ref"],
-            r["commit_sha"][:8],
-            r["url"],
-        )
-    console.print(tbl)
-
-
-@repo_group.command(name="search")
-@click.argument("query")
-@click.option("--limit", default=10, type=int, show_default=True)
-@click.option("--repo", default=None, help="Repo id/name/url filter.")
-@click.option("--path", "path_glob", default=None, help="Path glob filter, e.g. 'src/**/*.py'.")
-@click.option(
-    "--mode",
-    default="hybrid",
-    type=click.Choice(["hybrid", "vec", "bm25", "line"]),
-    show_default=True,
-)
-@click.option("--json", "as_json", is_flag=True)
-def repo_search_cmd(
-    query: str,
-    limit: int,
-    repo: str | None,
-    path_glob: str | None,
-    mode: str,
-    as_json: bool,
-) -> None:
-    """Search indexed repos by semantic chunks, keyword chunks, or exact lines."""
-    from memo.memory import Memory
-
-    mem = Memory(Config.from_env())
-    hits = mem.repo_search(query, limit=limit, repo=repo, path=path_glob, mode=mode)
-    if as_json:
-        click.echo(json.dumps([h.to_dict() for h in hits], ensure_ascii=False, indent=2))
-        return
-    if not hits:
-        console.print("[dim]no results[/dim]")
-        return
-    tbl = Table(show_lines=False, expand=True)
-    tbl.add_column("score", justify="right", width=7)
-    tbl.add_column("repo", width=16)
-    tbl.add_column("loc", overflow="fold")
-    tbl.add_column("text", overflow="fold")
-    for h in hits:
-        preview = " ".join((h.text or "").split())[:180]
-        tbl.add_row(
-            f"{h.score:.3f}" if h.score is not None else "—",
-            h.repo_name,
-            f"{h.path}:{h.line_start}-{h.line_end}",
-            preview,
-        )
-    console.print(tbl)
-
-
-@repo_group.command(name="get")
-@click.argument("repo")
-@click.argument("path")
-@click.option("--start", default=None, type=int, help="First line to return.")
-@click.option("--end", default=None, type=int, help="Last line to return.")
-@click.option("--json", "as_json", is_flag=True)
-def repo_get_cmd(
-    repo: str,
-    path: str,
-    start: int | None,
-    end: int | None,
-    as_json: bool,
-) -> None:
-    """Fetch one indexed repo file or line range."""
-    from memo.memory import Memory
-
-    mem = Memory(Config.from_env())
-    out = mem.repo_get_file(repo, path, start=start, end=end)
-    if out is None:
-        console.print(f"[red]not found:[/red] {repo} {path}")
-        sys.exit(1)
-    if as_json:
-        click.echo(json.dumps(out, ensure_ascii=False, indent=2))
-        return
-    console.print(Panel.fit(
-        out["text"],
-        title=f"{out['repo_name']}:{out['path']}:{out['start']}-{out['end']}",
-        border_style="cyan",
-    ))
-
-
-@repo_group.command(name="delete")
-@click.argument("repo")
-@click.option("--yes", is_flag=True, help="Skip confirmation.")
-@click.option("--keep-clone", is_flag=True, help="Only delete index rows; keep memo's managed clone.")
-def repo_delete_cmd(repo: str, yes: bool, keep_clone: bool) -> None:
-    """Delete one indexed repo."""
-    from memo.memory import Memory
-
-    mem = Memory(Config.from_env())
-    if not yes:
-        click.confirm(f"Delete indexed repo {repo!r}?", abort=True)
-    ok = mem.repo_delete(repo, remove_clone=not keep_clone)
-    console.print(f"[{'green' if ok else 'red'}]{'deleted' if ok else 'not found'}[/]: {repo}")
 
 
 @cli.command()
@@ -2580,49 +2182,6 @@ def lint(category: str | None, limit: int, as_json: bool) -> None:
             console.print(f"  · …and {n - limit} more")
 
 
-def _portable_backup(out_path: str | None) -> None:
-    """Snapshot memory dir + sqlite-vec DB + history DB into a zip.
-
-    Use before risky operations (mass migration, embedder swap, schema
-    change). The zip is portable: extract on another machine, set
-    `MEMO_VAULT_PATH` to the matching vault, run `memo restore <zip>`
-    to absorb everything back. Vault `.md` files are kept as the
-    storage of record so the backup is self-contained.
-    """
-    import datetime as _dt
-    import zipfile
-
-    cfg = Config.from_env()
-    cfg.ensure_dirs()
-    out = out_path or f"memo-backup-{_dt.datetime.now().strftime('%Y%m%d-%H%M%S')}.zip"
-    out_p = __import__("pathlib").Path(out).resolve()
-
-    n_md = 0
-    with zipfile.ZipFile(out_p, "w", zipfile.ZIP_DEFLATED) as zf:
-        # 1) Memory .md files (relative to memory_dir).
-        if cfg.memory_dir.is_dir():
-            for md in sorted(cfg.memory_dir.rglob("*.md")):
-                rel = md.relative_to(cfg.memory_dir)
-                zf.write(md, arcname=f"memory/{rel}")
-                n_md += 1
-        # 2) State DBs (vec + history). Stored at the root.
-        for db in (cfg.db_path, cfg.history_db):
-            if db.is_file():
-                zf.write(db, arcname=f"state/{db.name}")
-        # 3) Manifest with paths so restore can sanity-check.
-        manifest = {
-            "created": _dt.datetime.now().astimezone().isoformat(timespec="seconds"),
-            "data_dir": str(cfg.data_dir),
-            "vault_path": str(cfg.vault_path) if cfg.vault_path else None,
-            "embedder_model": cfg.embedder_model,
-            "embedder_dims": cfg.embedder_dims,
-            "memo_version": __import__("memo").__version__,
-            "n_md": n_md,
-        }
-        zf.writestr("manifest.json", json.dumps(manifest, indent=2))
-
-    size_kb = out_p.stat().st_size // 1024
-    console.print(f"[green]✓[/green] backup: {out_p} ({n_md} memorias, {size_kb} KB)")
 
 
 @cli.command()
@@ -3524,339 +3083,10 @@ def recall_hook() -> None:
 # ── Recall daemon — persistent socket server for low-latency recall ──────────
 
 
-@cli.group(name="recall-daemon")
-def recall_daemon_group() -> None:
-    """Manage the persistent recall daemon (fast socket-based recall).
-
-    The daemon keeps the MLX embedder in RAM so recall-hook answers in
-    <200 ms instead of 1-2 s per prompt (cold Python + MLX load).
-
-    Subcommands: start, stop, status, (internal) _serve.
-    """
 
 
-@recall_daemon_group.command(name="start")
-def recall_daemon_start() -> None:
-    """Start the recall daemon in the background."""
-    import subprocess as _subprocess
-
-    from memo.config import Config
-    from memo.recall_server import _is_pid_alive, _read_pid, _socket_path
-
-    cfg = Config.from_env()
-    pid = _read_pid(cfg.state_dir)
-    if pid is not None and _is_pid_alive(pid):
-        console.print(f"[dim]recall daemon already running (pid={pid})[/dim]")
-        return
-
-    log_dir = Path.home() / "Library" / "Logs" / "memo"
-    log_dir.mkdir(parents=True, exist_ok=True)
-    log_file = log_dir / "recall-daemon.log"
-
-    env = os.environ.copy()
-    env["MEMO_NONINTERACTIVE"] = "1"
-
-    with open(log_file, "a") as lf:
-        proc = _subprocess.Popen(
-            [sys.executable, "-m", "memo.cli", "recall-daemon", "_serve"],
-            stdout=lf,
-            stderr=lf,
-            stdin=_subprocess.DEVNULL,  # don't inherit hook stdin pipe
-            start_new_session=True,
-            env=env,
-        )
-
-    # Wait briefly for the socket to appear
-    sock_path = _socket_path(cfg.state_dir)
-    for _ in range(20):
-        time.sleep(0.1)
-        if sock_path.exists():
-            break
-
-    click.echo(f"recall daemon started (pid={proc.pid})", err=True)
 
 
-@recall_daemon_group.command(name="stop")
-def recall_daemon_stop() -> None:
-    """Stop the recall daemon (sends SIGTERM)."""
-    import signal as _signal
-
-    from memo.config import Config
-    from memo.recall_server import _cleanup, _is_pid_alive, _read_pid
-
-    cfg = Config.from_env()
-    pid = _read_pid(cfg.state_dir)
-    if pid is None:
-        console.print("[dim]recall daemon not running (no PID file)[/dim]")
-        return
-    if not _is_pid_alive(pid):
-        console.print("[dim]recall daemon not running (stale PID file)[/dim]")
-        _cleanup(cfg.state_dir)
-        return
-    try:
-        os.kill(pid, _signal.SIGTERM)
-        console.print(f"recall daemon stopped (pid={pid})")
-    except ProcessLookupError:
-        console.print("[dim]recall daemon already gone[/dim]")
-    _cleanup(cfg.state_dir)
-
-
-@recall_daemon_group.command(name="status")
-def recall_daemon_status() -> None:
-    """Print whether the recall daemon is running."""
-    from memo.config import Config
-    from memo.recall_server import _is_pid_alive, _read_pid, _socket_path
-
-    cfg = Config.from_env()
-    pid = _read_pid(cfg.state_dir)
-    sock = _socket_path(cfg.state_dir)
-
-    if pid is not None and _is_pid_alive(pid):
-        console.print(f"[green]running[/green] pid={pid}  socket={sock}")
-    else:
-        console.print(f"[red]stopped[/red]  socket={sock}")
-
-
-@recall_daemon_group.command(name="_serve", hidden=True)
-def recall_daemon_serve() -> None:
-    """Internal: run the daemon in the foreground (called by 'start')."""
-    from memo.recall_server import run_server
-    run_server()
-
-
-# ── Embed daemon — observability over the shared embedder sidecar ────────────
-
-
-@cli.group(name="embed-daemon")
-def embed_daemon_group() -> None:
-    """Inspect the recall daemon's shared embedder sidecar.
-
-    The recall daemon doubles as a shared MLX embedder so peers
-    (synapse, memflow, other memo CLIs) can reuse one warm model.
-    These commands surface its metrics without opening the socket
-    by hand. See `memo.embedder_client` for the in-process client.
-    """
-
-
-@embed_daemon_group.command(name="status")
-def embed_daemon_status() -> None:
-    """Show whether the daemon is alive plus model + uptime."""
-    from memo.embedder_client import status as _status
-
-    info = _status()
-    if info is None:
-        console.print("[red]daemon: not running[/red]")
-        sys.exit(1)
-    uptime = info.get("uptime_s")
-    uptime_str = f"{uptime}s" if isinstance(uptime, int) else "?"
-    console.print(
-        f"[green]daemon: running[/green] "
-        f"model={info.get('model', '?')} dims={info.get('dims', '?')} "
-        f"uptime={uptime_str}"
-    )
-
-
-@embed_daemon_group.command(name="stats")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def embed_daemon_stats(as_json: bool) -> None:
-    """Per-op request counters + latency p50/p95/p99."""
-    from memo.embedder_client import stats as _stats
-
-    info = _stats()
-    if info is None:
-        console.print("[red]daemon: not running[/red]")
-        sys.exit(1)
-
-    if as_json:
-        click.echo(json.dumps(info, indent=2))
-        return
-
-    uptime = info.get("uptime_s")
-    uptime_str = f"{uptime}s" if isinstance(uptime, int) else "?"
-    console.print(
-        f"[bold]Embedder daemon stats[/bold] "
-        f"model={info.get('model', '?')} dims={info.get('dims', '?')} "
-        f"uptime={uptime_str}"
-    )
-    ops = info.get("ops") or {}
-    if not ops:
-        console.print("[dim]no requests recorded yet[/dim]")
-        return
-
-    table = Table(show_header=True, header_style="bold")
-    table.add_column("op")
-    table.add_column("count", justify="right")
-    table.add_column("errors", justify="right")
-    table.add_column("samples", justify="right")
-    table.add_column("p50 ms", justify="right")
-    table.add_column("p95 ms", justify="right")
-    table.add_column("p99 ms", justify="right")
-    for op, row in sorted(ops.items()):
-        def _fmt(v: float | None) -> str:
-            return "-" if v is None else f"{v:.1f}"
-        table.add_row(
-            op,
-            str(row.get("count", 0)),
-            str(row.get("errors", 0)),
-            str(row.get("samples", 0)),
-            _fmt(row.get("p50_ms")),
-            _fmt(row.get("p95_ms")),
-            _fmt(row.get("p99_ms")),
-        )
-    console.print(table)
-
-
-@cli.group(name="as-of")
-def as_of_group() -> None:
-    """Time-machine — query the corpus as it existed at any past date.
-
-    Subcommands: `search`, `ask`, `list`. All take `--date YYYY-MM-DD`
-    (or a full ISO timestamp). The snapshot is reconstructed by
-    replaying `history.db` events in reverse from "now".
-    """
-    pass
-
-
-def _parse_as_of_date(s: str) -> str:
-    """Accept date-only (`2026-03-01`) or full ISO. Return ISO with
-    a stable noon-UTC anchor for date-only inputs."""
-    from datetime import UTC
-    from datetime import datetime as _dt
-    s = s.strip()
-    if len(s) == 10:  # YYYY-MM-DD
-        return f"{s}T23:59:59+00:00"  # end-of-day to be inclusive
-    try:
-        dt = _dt.fromisoformat(s.rstrip("Z"))
-        if dt.tzinfo is None:
-            dt = dt.replace(tzinfo=UTC)
-        return dt.isoformat()
-    except ValueError as exc:
-        raise click.ClickException(
-            f"Could not parse --date {s!r}. Use YYYY-MM-DD or ISO 8601.",
-        ) from exc
-
-
-@as_of_group.command(name="search")
-@click.argument("query")
-@click.option("--date", "as_of", required=True, help="YYYY-MM-DD or full ISO 8601.")
-@click.option("--limit", default=10, type=int, show_default=True)
-@click.option("--type", "type_", default=None, help="Filter by record type.")
-@click.option("--mode", default="hybrid",
-              type=click.Choice(["hybrid", "vec", "bm25"]), show_default=True)
-@click.option("--json", "as_json", is_flag=True)
-def as_of_search(
-    query: str, as_of: str, limit: int, type_: str | None, mode: str, as_json: bool,
-) -> None:
-    """Search the corpus as it existed on a past date."""
-    from memo.memory import Memory
-    from memo.time_machine import reconstruct
-
-    mem = Memory(Config.from_env())
-    snap = reconstruct(mem, as_of=_parse_as_of_date(as_of))
-    hits = snap.search(query, limit=limit, mode=mode)
-    if type_:
-        hits = [h for h in hits if h.type == type_]
-
-    if as_json:
-        click.echo(json.dumps(
-            {
-                "as_of": snap.as_of.isoformat(),
-                "snapshot_size": len(snap),
-                "results": [h.to_dict() for h in hits],
-            },
-            ensure_ascii=False, indent=2,
-        ))
-        return
-
-    if not hits:
-        console.print(f"[dim]no results in snapshot @ {snap.as_of.date().isoformat()}[/dim]")
-        return
-    tbl = Table(show_lines=False, expand=True,
-                title=f"snapshot @ {snap.as_of.date().isoformat()} · {len(snap)} memorias existían")
-    tbl.add_column("score", justify="right", width=6)
-    tbl.add_column("type", width=10)
-    tbl.add_column("title", overflow="fold")
-    tbl.add_column("tags", overflow="fold")
-    for h in hits:
-        tbl.add_row(
-            f"{h.score:.3f}" if h.score is not None else "—",
-            h.type,
-            h.title,
-            ", ".join(h.tags) or "—",
-        )
-    console.print(tbl)
-
-
-@as_of_group.command(name="ask")
-@click.argument("question")
-@click.option("--date", "as_of", required=True, help="YYYY-MM-DD or full ISO 8601.")
-@click.option("--k", default=5, type=int, show_default=True)
-@click.option("--json", "as_json", is_flag=True)
-def as_of_ask(question: str, as_of: str, k: int, as_json: bool) -> None:
-    """RAG question against a past snapshot of the corpus."""
-    from memo.memory import Memory
-    from memo.time_machine import reconstruct
-
-    mem = Memory(Config.from_env())
-    snap = reconstruct(mem, as_of=_parse_as_of_date(as_of))
-    out = snap.ask(question, k=k)
-
-    if as_json:
-        click.echo(json.dumps(out, ensure_ascii=False, indent=2))
-        return
-
-    console.print(Panel.fit(
-        out["answer"] or "[dim](sin respuesta)[/dim]",
-        title=f"✓ as-of {snap.as_of.date().isoformat()} ({len(snap)} memorias in scope)",
-        border_style="magenta",
-    ))
-    if out.get("sources"):
-        console.print("\n[dim]sources:[/dim]")
-        for s in out["sources"]:
-            console.print(f"  [bold]{s['id_short']}[/bold]  {s['title']}  [dim]({s['type']})[/dim]")
-
-
-@as_of_group.command(name="list")
-@click.option("--date", "as_of", required=True, help="YYYY-MM-DD or full ISO 8601.")
-@click.option("--type", "type_", default=None)
-@click.option("--limit", default=20, type=int, show_default=True)
-@click.option("--json", "as_json", is_flag=True)
-def as_of_list(as_of: str, type_: str | None, limit: int, as_json: bool) -> None:
-    """List memorias that existed in a past snapshot (most-recent first)."""
-    from memo.memory import Memory
-    from memo.time_machine import reconstruct
-
-    mem = Memory(Config.from_env())
-    snap = reconstruct(mem, as_of=_parse_as_of_date(as_of))
-    rows = snap.list(type_=type_)[:limit]
-
-    if as_json:
-        click.echo(json.dumps(
-            {
-                "as_of": snap.as_of.isoformat(),
-                "snapshot_size": len(snap),
-                "records": [
-                    {"id": r.id, "title": r.title, "type": r.type, "tags": r.tags,
-                     "updated": r.updated}
-                    for r in rows
-                ],
-            },
-            ensure_ascii=False, indent=2,
-        ))
-        return
-
-    if not rows:
-        console.print(f"[dim]empty snapshot @ {snap.as_of.date().isoformat()}[/dim]")
-        return
-    tbl = Table(show_lines=False, expand=True,
-                title=f"snapshot @ {snap.as_of.date().isoformat()} · {len(snap)} memorias")
-    tbl.add_column("id", width=10)
-    tbl.add_column("type", width=10)
-    tbl.add_column("title", overflow="fold")
-    tbl.add_column("updated", width=12)
-    for r in rows:
-        tbl.add_row(r.id[:8], r.type, r.title, (r.updated or "—")[:10])
-    console.print(tbl)
 
 
 @cli.command(name="diff")
@@ -5117,232 +4347,6 @@ def prewarm(download_all: bool) -> None:
 # query pattern — looked up by recency, never by semantic similarity).
 
 
-@cli.group(name="session")
-def session_group() -> None:
-    """Internal session-snapshot ops — hook targets, not user-facing.
-
-    User-facing entry is `memo resume` (list / inspect). This group
-    holds the wiring the hooks call: `checkpoint` (Stop hook), `recent`
-    (SessionStart additionalContext), and `prune` (LRU cleanup). Stays
-    namespaced so `memo --help` doesn't surface plumbing as if it were
-    everyday CLI.
-    """
-
-
-@session_group.command(name="checkpoint")
-@click.option("--session-id", default=None, help="Override session_id (default: read from stdin payload).")
-@click.option("--cwd", default=None, help="Override cwd (default: read from stdin payload, fallback os.getcwd).")
-@click.option("--transcript-path", default=None, help="Override transcript path.")
-@click.option("--lru-cap", default=50, type=int, show_default=True)
-@click.option("--json", "as_json", is_flag=True, help="Print the persisted snapshot as JSON.")
-def session_checkpoint(
-    session_id: str | None, cwd: str | None,
-    transcript_path: str | None, lru_cap: int, as_json: bool,
-) -> None:
-    """Stop hook entrypoint — upsert a session snapshot from stdin JSON.
-
-    Reads the Stop hook payload from stdin (Claude Code passes
-    `{"session_id", "transcript_path", "cwd", ...}`). Falls back to
-    flags/cwd if stdin is empty (lets you run it manually for testing).
-
-    Always exits 0 — like the other hooks, a checkpoint failure must
-    not block Claude Code. On any exception we swallow + print `{}`.
-    """
-    import json as _json
-    import os as _os
-    import sys as _sys
-
-    if _os.environ.get("MEMO_SESSION_DISABLE") == "1":
-        if as_json:
-            click.echo("{}")
-        sys.exit(0)
-
-    payload: dict[str, Any] = {}
-    # Stdin is a TTY when run interactively → don't block on read.
-    if not _sys.stdin.isatty():
-        try:
-            raw = _sys.stdin.read()
-            if raw.strip():
-                payload = _json.loads(raw)
-        except _json.JSONDecodeError:
-            payload = {}
-
-    sid = session_id or payload.get("session_id")
-    cwd_resolved = cwd or payload.get("cwd") or _os.getcwd()
-    transcript = transcript_path or payload.get("transcript_path")
-    prompt_text = payload.get("prompt")  # present in UserPromptSubmit events
-
-    if not sid:
-        # Without a session_id we can't key the snapshot. Fail silently
-        # so the hook still exits 0.
-        if as_json:
-            click.echo("{}")
-        sys.exit(0)
-
-    try:
-        from memo.session import checkpoint as _checkpoint
-
-        cfg = Config.from_env()
-        cfg.ensure_dirs()
-        snap = _checkpoint(
-            cfg.state_dir,
-            session_id=sid,
-            cwd=cwd_resolved,
-            transcript_path=transcript,
-            prompt=prompt_text,
-            lru_cap=lru_cap,
-        )
-    except Exception as exc:
-        if _os.environ.get("MEMO_SESSION_DEBUG") == "1":
-            print(f"# memo session checkpoint failed: {exc}", file=_sys.stderr)
-        if as_json:
-            click.echo("{}")
-        sys.exit(0)
-
-    if as_json:
-        click.echo(_json.dumps(snap, ensure_ascii=False, indent=2))
-
-
-@session_group.command(name="autosave")
-@click.option("--threshold-kb", default=1024, type=int, show_default=True,
-              help="Transcript size (KB) that triggers an autosave.")
-@click.option("--cooldown", default=300, type=int, show_default=True,
-              help="Minimum seconds between autosaves for the same session.")
-def session_autosave(threshold_kb: int, cooldown: int) -> None:
-    """UserPromptSubmit hook — proactive save when context approaches limits.
-
-    Stats the transcript file (fast, O(1)). If it exceeds threshold_kb
-    and the per-session cooldown has elapsed, spawns ``memo capture-stop``
-    in a detached background process and emits an additionalContext warning.
-
-    Always exits 0 — never blocks the prompt.
-    """
-    import json as _json
-    import os as _os
-    import subprocess as _sp
-    import sys as _sys
-
-    if _os.environ.get("MEMO_SESSION_DISABLE") == "1":
-        print("{}")
-        _sys.exit(0)
-
-    payload: dict[str, Any] = {}
-    if not _sys.stdin.isatty():
-        try:
-            raw = _sys.stdin.read()
-            if raw.strip():
-                payload = _json.loads(raw)
-        except _json.JSONDecodeError:
-            payload = {}
-
-    sid = payload.get("session_id")
-    transcript = payload.get("transcript_path")
-    cwd = payload.get("cwd") or _os.getcwd()
-
-    if not sid or not transcript:
-        print("{}")
-        _sys.exit(0)
-
-    try:
-        from memo.session import check_autosave, mark_autosaved
-        cfg = Config.from_env()
-        should_save, size_kb = check_autosave(
-            cfg.state_dir,
-            session_id=sid,
-            transcript_path=transcript,
-            threshold_kb=threshold_kb,
-            cooldown_secs=cooldown,
-        )
-    except Exception as exc:
-        if _os.environ.get("MEMO_SESSION_DEBUG") == "1":
-            print(f"# session autosave check failed: {exc}", file=_sys.stderr)
-        print("{}")
-        _sys.exit(0)
-
-    if not should_save:
-        print("{}")
-        _sys.exit(0)
-
-    # Spawn capture-stop detached, passing the hook payload via stdin.
-    try:
-        env = {**_os.environ, "MEMO_NONINTERACTIVE": "1"}
-        capture_payload = _json.dumps({
-            "session_id": sid,
-            "transcript_path": transcript,
-            "cwd": cwd,
-        }).encode()
-        proc = _sp.Popen(
-            ["memo", "capture-stop"],
-            stdin=_sp.PIPE,
-            stdout=_sp.DEVNULL,
-            stderr=_sp.DEVNULL,
-            start_new_session=True,
-            env=env,
-            cwd=cwd,
-        )
-        if proc.stdin is not None:
-            proc.stdin.write(capture_payload)
-            proc.stdin.close()
-        mark_autosaved(cfg.state_dir, sid)
-    except Exception as exc:
-        if _os.environ.get("MEMO_SESSION_DEBUG") == "1":
-            print(f"# session autosave spawn failed: {exc}", file=_sys.stderr)
-        print("{}")
-        _sys.exit(0)
-
-    output = {
-        "hookSpecificOutput": {
-            "hookEventName": "UserPromptSubmit",
-            "additionalContext": (
-                "## ⚠️ Snapshot automático\n\n"
-                f"El transcript de esta sesión supera los {size_kb} KB "
-                f"(umbral: {threshold_kb} KB) — contexto próximo al límite. "
-                "Lancé `memo capture-stop` en segundo plano para preservar "
-                "los insights más importantes. Podés continuar con normalidad; "
-                "si hay compactación automática, lo esencial ya está en memo.\n"
-            ),
-        }
-    }
-    print(_json.dumps(output, ensure_ascii=False))
-
-
-@session_group.command(name="refresh-summary")
-def session_refresh_summary() -> None:
-    """Stop hook entrypoint — generate/update running_summary for the active session.
-
-    Reads the Stop hook payload from stdin, runs the 3B LLM helper to
-    summarize the session arc, and stores it in the snapshot. Throttled:
-    skips if fewer than 3 new turns since the last summary. Always exits 0.
-    """
-    import json as _json
-    import os as _os
-    import sys as _sys
-
-    if _os.environ.get("MEMO_SESSION_DISABLE") == "1":
-        _sys.exit(0)
-
-    payload: dict[str, Any] = {}
-    if not _sys.stdin.isatty():
-        try:
-            raw = _sys.stdin.read()
-            if raw.strip():
-                payload = _json.loads(raw)
-        except _json.JSONDecodeError:
-            payload = {}
-
-    sid = payload.get("session_id")
-    if not sid:
-        _sys.exit(0)
-
-    try:
-        from memo.session import refresh_summary as _refresh_summary
-        cfg = Config.from_env()
-        cfg.ensure_dirs()
-        _refresh_summary(cfg.state_dir, sid)
-    except Exception as exc:
-        if _os.environ.get("MEMO_SESSION_DEBUG") == "1":
-            print(f"# memo session refresh-summary failed: {exc}", file=_sys.stderr)
-    _sys.exit(0)
 
 
 @cli.command(name="resume")
@@ -5828,2623 +4832,47 @@ def install_shell_wrapper(
         )
 
 
-# -- temporal reasoning commands ----------------------------------------------
 
 
-@cli.group(name="temporal")
-def temporal_group() -> None:
-    """Analyze temporal patterns and contradictions in memories."""
-    pass
 
 
-@temporal_group.command(name="contradictions")
-@click.argument("entity")
-@click.option("--type", "entity_type", help="Filter by entity type from graph")
-@click.option("--confidence", "min_confidence", type=float, default=0.7,
-              help="Minimum confidence threshold (default: 0.7)")
-@click.option("--max-pairs", type=int, default=20,
-              help="Maximum number of pairs to analyze (default: 20)")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def temporal_contradictions(
-    entity: str, entity_type: str | None, min_confidence: float,
-    max_pairs: int, as_json: bool,
-) -> None:
-    """Detect contradictions among memorias mentioning a specific entity.
-
-    Example: memo temporal contradictions mlx
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    contradictions = mem.temporal.detect_entity_contradictions(
-        entity_name=entity,
-        entity_type=entity_type,
-        confidence_threshold=min_confidence,
-        max_pairs=max_pairs,
-    )
-
-    if as_json:
-        click.echo(json.dumps([c.__dict__ for c in contradictions], indent=2))
-        return
-
-    if not contradictions:
-        console.print(f"[green]No contradictions found for entity '{entity}'[/green]")
-        return
-
-    table = Table(title=f"Contradictions for '{entity}'")
-    table.add_column("ID A", style="cyan")
-    table.add_column("ID B", style="cyan")
-    table.add_column("Title A", style="yellow")
-    table.add_column("Title B", style="yellow")
-    table.add_column("Relationship", style="magenta")
-    table.add_column("Confidence", style="green")
-    table.add_column("Rationale")
-
-    for c in contradictions:
-        table.add_row(
-            c.memoria_id_a[:8],
-            c.memoria_id_b[:8],
-            c.title_a[:40],
-            c.title_b[:40],
-            c.relationship,
-            f"{c.confidence:.2f}",
-            c.rationale,
-        )
-
-    console.print(table)
-
-
-@temporal_group.command(name="timeline")
-@click.argument("entity")
-@click.option("--type", "entity_type", help="Filter by entity type from graph")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def temporal_timeline(entity: str, entity_type: str | None, as_json: bool) -> None:
-    """Build a chronological timeline of all memorias mentioning an entity.
-
-    Example: memo temporal timeline mlx
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    timeline = mem.temporal.build_entity_timeline(
-        entity_name=entity,
-        entity_type=entity_type,
-    )
-
-    if timeline is None:
-        console.print(f"[yellow]No memorias found for entity '{entity}'[/yellow]")
-        return
-
-    if as_json:
-        click.echo(json.dumps({
-            "entity_name": timeline.entity_name,
-            "entity_type": timeline.entity_type,
-            "first_seen": timeline.first_seen,
-            "last_seen": timeline.last_seen,
-            "events": [e.__dict__ for e in timeline.events],
-        }, indent=2))
-        return
-
-    console.print(f"[bold]Timeline for '{entity}' ({timeline.entity_type})[/bold]")
-    console.print(f"First seen: {timeline.first_seen}")
-    console.print(f"Last seen: {timeline.last_seen}")
-    console.print()
-
-    for event in timeline.events:
-        console.print(f"[cyan]{event.date}[/cyan] [dim][{event.memoria_id[:8]}][/dim]")
-        console.print(f"  [yellow]{event.title}[/yellow] ({event.type})")
-        console.print(f"  {event.snippet}")
-        console.print()
-
-
-@temporal_group.command(name="stale")
-@click.option("--days", type=int, default=180,
-              help="Days since last update to consider stale (default: 180)")
-@click.option("--min-access", "min_access_count", type=int, default=0,
-              help="Minimum access count to exclude (default: 0)")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def temporal_stale(days: int, min_access_count: int, as_json: bool) -> None:
-    """Find memorias that may be stale based on age and lack of access.
-
-    Example: memo temporal stale --days 90
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    stale = mem.temporal.detect_stale_memorias(
-        days_threshold=days,
-        min_access_count=min_access_count,
-    )
-
-    if as_json:
-        click.echo(json.dumps(stale, indent=2))
-        return
-
-    if not stale:
-        console.print(f"[green]No stale memorias found (threshold: {days} days)[/green]")
-        return
-
-    table = Table(title=f"Potentially Stale Memorias (>{days} days)")
-    table.add_column("ID", style="cyan")
-    table.add_column("Title", style="yellow")
-    table.add_column("Type", style="magenta")
-    table.add_column("Updated", style="dim")
-    table.add_column("Days Old", style="red")
-    table.add_column("Access Count", style="green")
-
-    for item in stale[:50]:  # Cap display
-        table.add_row(
-            item["id"][:8],
-            item["title"][:40],
-            item["type"],
-            item["updated"][:10],
-            str(item["days_since_update"]),
-            str(item["access_count"]),
-        )
-
-    console.print(table)
-    if len(stale) > 50:
-        console.print(f"[dim]...and {len(stale) - 50} more[/dim]")
-
-
-@temporal_group.command(name="patterns")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def temporal_patterns(as_json: bool) -> None:
-    """Analyze high-level temporal patterns across the entire corpus.
-
-    Example: memo temporal patterns
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    patterns = mem.temporal.detect_temporal_patterns()
-
-    if as_json:
-        click.echo(json.dumps(patterns, indent=2))
-        return
-
-    console.print("[bold]Temporal Patterns[/bold]")
-    console.print()
-
-    # Memorias per month
-    console.print("[yellow]Memorias per month:[/yellow]")
-    for month, count in list(patterns["memorias_per_month"].items())[-12:]:
-        console.print(f"  {month}: {count}")
-    console.print()
-
-    # Most active entities
-    console.print("[yellow]Most active entities:[/yellow]")
-    for entity, count in list(patterns["most_active_entities"].items())[:10]:
-        console.print(f"  {entity}: {count} mentions")
-
-
-
-
-# -- advanced consolidation commands -------------------------------------------
-
-
-@cli.group(name="consolidate")
-def consolidate_group() -> None:
-    """Advanced consolidation with intelligent merge and archival."""
-    pass
-
-
-@consolidate_group.command(name="propose")
-@click.option("--threshold", type=float, default=0.85,
-              help="Cosine similarity threshold (default: 0.85)")
-@click.option("--max-clusters", type=int, default=20,
-              help="Maximum clusters to process (default: 20)")
-@click.option("--type", "type_", help="Filter by memoria type")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def consolidate_propose(
-    threshold: float, max_clusters: int, type_: str | None, as_json: bool,
-) -> None:
-    """Detect clusters and propose merge strategies (read-only).
-
-    Example: memo consolidate propose --threshold 0.9
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    result = mem.consolidator.consolidate_all(
-        threshold=threshold,
-        max_clusters=max_clusters,
-        type_=type_,
-        auto_apply=False,
-        dry_run=True,
-    )
-
-    if as_json:
-        click.echo(json.dumps(result, indent=2))
-        return
-
-    clusters = result.get("clusters", [])
-    proposals = result.get("proposals", [])
-
-    console.print(f"[bold]Detected {len(clusters)} clusters[/bold]")
-    console.print(f"[yellow]Generated {len(proposals)} merge proposals[/yellow]")
-    console.print()
-
-    if not proposals:
-        console.print("[green]No mergeable clusters found[/green]")
-        return
-
-    for i, p in enumerate(proposals[:10], 1):
-        console.print(f"[cyan]{i}. Cluster {p['cluster_id']}[/cyan]")
-        console.print(f"   Strategy: {p['merge_strategy']}")
-        console.print(f"   Rationale: {p['rationale']}")
-        console.print(f"   Memorias to merge: {len(p['memoria_ids'])}")
-        console.print()
-
-    if len(proposals) > 10:
-        console.print(f"[dim]...and {len(proposals) - 10} more proposals[/dim]")
-
-
-@consolidate_group.command(name="apply")
-@click.option("--threshold", type=float, default=0.85,
-              help="Cosine similarity threshold (default: 0.85)")
-@click.option("--max-clusters", type=int, default=20,
-              help="Maximum clusters to process (default: 20)")
-@click.option("--type", "type_", help="Filter by memoria type")
-@click.option("--dry-run", is_flag=True,
-              help="Show what would happen without applying changes")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-@click.confirmation_option(prompt="This will merge memorias and archive old ones. Continue?")
-def consolidate_apply(
-    threshold: float, max_clusters: int, type_: str | None,
-    dry_run: bool, as_json: bool,
-) -> None:
-    """Apply merge proposals to consolidate the corpus.
-
-    Example: memo consolidate apply --dry-run
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    result = mem.consolidator.consolidate_all(
-        threshold=threshold,
-        max_clusters=max_clusters,
-        type_=type_,
-        auto_apply=True,
-        dry_run=dry_run,
-    )
-
-    if as_json:
-        click.echo(json.dumps(result, indent=2))
-        return
-
-    result.get("proposals", [])
-    results = result.get("results", [])
-
-    if dry_run:
-        console.print("[yellow]Dry run mode - no changes applied[/yellow]")
-        console.print()
-
-    console.print(f"[bold]Processed {len(results)} consolidations[/bold]")
-    console.print()
-
-    merged_count = sum(1 for r in results if r.get("merged_id"))
-    archived_count = sum(len(r.get("archived_ids", [])) for r in results)
-    skipped_count = sum(len(r.get("skipped_ids", [])) for r in results)
-
-    console.print(f"[green]✓ Merged {merged_count} memorias[/green]")
-    console.print(f"[yellow]↻ Archived {archived_count} old versions[/yellow]")
-    console.print(f"[dim]⊘ Skipped {skipped_count} (conflicts)[/dim]")
-
-    for r in results[:5]:
-        console.print(f"  {r.get('summary', '')}")
-
-    if len(results) > 5:
-        console.print(f"[dim]...and {len(results) - 5} more[/dim]")
-
-
-@consolidate_group.command(name="list-archived")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def consolidate_list_archived(as_json: bool) -> None:
-    """List all archived memorias.
-
-    Example: memo consolidate list-archived
-    """
-    cfg = Config.from_env()
-    archival_dir = cfg.memory_dir / "archived"
-
-    if not archival_dir.is_dir():
-        console.print("[dim]No archived memorias found[/dim]")
-        return
-
-    archived_files = list(archival_dir.glob("*.md"))
-
-    if as_json:
-        archived_data = []
-        for f in archived_files:
-            import frontmatter
-            post = frontmatter.loads(f.read_text(encoding="utf-8"))
-            archived_data.append({
-                "id": f.stem,
-                "title": post.get("title", ""),
-                "archived_for": post.get("archived_for", ""),
-                "archived_at": post.get("archived_at", ""),
-            })
-        click.echo(json.dumps(archived_data, indent=2))
-        return
-
-    table = Table(title="Archived Memorias")
-    table.add_column("ID", style="cyan")
-    table.add_column("Title", style="yellow")
-    table.add_column("Archived For", style="green")
-    table.add_column("Archived At", style="dim")
-
-    for f in archived_files[:50]:
-        import frontmatter
-        post = frontmatter.loads(f.read_text(encoding="utf-8"))
-        table.add_row(
-            f.stem[:8],
-            str(post.get("title") or "")[:40],
-            str(post.get("archived_for") or "")[:8],
-            str(post.get("archived_at") or "")[:10],
-        )
-
-    console.print(table)
-    if len(archived_files) > 50:
-        console.print(f"[dim]...and {len(archived_files) - 50} more[/dim]")
 
 
 # -- graph navigation commands ------------------------------------------------
 
 
-# -- contextual recall commands -----------------------------------------------
 
 
-@cli.group(name="contextual")
-def contextual_group() -> None:
-    """Contextual recall with conversation history and preference learning."""
-    pass
 
 
-@contextual_group.command(name="search")
-@click.argument("query")
-@click.option("--limit", type=int, default=10,
-              help="Max results (default: 10)")
-@click.option("--mode", type=click.Choice(["vec", "bm25", "hybrid"]), default="hybrid",
-              help="Search mode (default: hybrid)")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def contextual_search(query: str, limit: int, mode: str, as_json: bool) -> None:
-    """Search with contextual re-ranking based on conversation history.
 
-    Example: memo contextual search "MLX performance"
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
 
-    results = mem.contextual.search_with_context(
-        query=query,
-        limit=limit,
-        mode=mode,
-    )
 
-    if as_json:
-        click.echo(json.dumps([r.__dict__ for r in results], indent=2))
-        return
 
-    if not results:
-        console.print("[yellow]No results found[/yellow]")
-        return
 
-    table = Table(title=f"Contextual Search Results for '{query}'")
-    table.add_column("ID", style="cyan")
-    table.add_column("Title", style="yellow")
-    table.add_column("Original Score", style="dim")
-    table.add_column("Contextual Score", style="green")
-    table.add_column("Boost Factors", style="magenta")
 
-    for r in results[:20]:
-        boosts = ", ".join(f"{k}={v:.2f}" for k, v in r.boost_factors.items())
-        table.add_row(
-            r.memoria_id[:8],
-            r.title[:40],
-            f"{r.original_score:.3f}",
-            f"{r.contextual_score:.3f}",
-            boosts or "—",
-        )
 
-    console.print(table)
-    if len(results) > 20:
-        console.print(f"[dim]...and {len(results) - 20} more[/dim]")
 
 
-@contextual_group.command(name="record-search")
-@click.argument("query")
-@click.argument("memoria_ids", nargs=-1, required=True)
-def contextual_record_search(query: str, memoria_ids: tuple[str, ...]) -> None:
-    """Record a search in the conversation history for learning.
 
-    Example: memo contextual record-search "MLX" abc123 def456
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
 
-    mem.contextual.record_search(query, list(memoria_ids))
-    console.print(f"[green]Recorded search with {len(memoria_ids)} recalled memorias[/green]")
 
 
-@contextual_group.command(name="record-click")
-@click.argument("memoria_id")
-def contextual_record_click(memoria_id: str) -> None:
-    """Record that the user clicked/viewed a memoria (for preference learning).
 
-    Example: memo contextual record-click abc123
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
 
-    mem.contextual.record_click(memoria_id)
-    console.print(f"[green]Recorded click for memoria {memoria_id[:8]}[/green]")
 
 
-@contextual_group.command(name="preferences")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def contextual_preferences(as_json: bool) -> None:
-    """Show learned user preferences for memory recall.
 
-    Example: memo contextual preferences
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
 
-    prefs = mem.contextual.context.get_preferences()
 
-    if as_json:
-        click.echo(json.dumps(prefs.__dict__, indent=2))
-        return
 
-    console.print("[bold]User Preferences[/bold]")
-    console.print()
 
-    console.print(f"[yellow]Recency Weight:[/yellow] {prefs.recency_weight:.2f}")
-    console.print(f"[yellow]Diversity Weight:[/yellow] {prefs.diversity_weight:.2f}")
-    console.print(f"[yellow]Last Updated:[/yellow] {prefs.last_updated or 'Never'}")
-    console.print()
 
-    console.print("[yellow]Preferred Memory Types:[/yellow]")
-    if prefs.preferred_types:
-        for type_, score in sorted(prefs.preferred_types.items(), key=lambda x: x[1], reverse=True):
-            console.print(f"  {type_}: {score:.2f}")
-    else:
-        console.print("  [dim]No preferences learned yet[/dim]")
-    console.print()
 
-    console.print("[yellow]Preferred Entities:[/yellow]")
-    if prefs.preferred_entities:
-        for entity, score in sorted(prefs.preferred_entities.items(), key=lambda x: x[1], reverse=True)[:10]:
-            console.print(f"  {entity}: {score:.2f}")
-        if len(prefs.preferred_entities) > 10:
-            console.print(f"  [dim]...and {len(prefs.preferred_entities) - 10} more[/dim]")
-    else:
-        console.print("  [dim]No preferences learned yet[/dim]")
 
 
-@contextual_group.command(name="history")
-@click.option("--limit", type=int, default=10,
-              help="Number of recent prompts to show (default: 10)")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def contextual_history(limit: int, as_json: bool) -> None:
-    """Show recent conversation history used for contextual recall.
 
-    Example: memo contextual history --limit 5
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
 
-    history = mem.contextual.context.get_recent_context(n=limit)
-
-    if as_json:
-        click.echo(json.dumps([c.__dict__ for c in history], indent=2))
-        return
-
-    if not history:
-        console.print("[dim]No conversation history yet[/dim]")
-        return
-
-    console.print(f"[bold]Recent {len(history)} Prompts[/bold]")
-    console.print()
-
-    for i, ctx in enumerate(history, 1):
-        console.print(f"[cyan]{i}. {ctx.timestamp}[/cyan]")
-        console.print(f"   Prompt: {ctx.prompt[:80]}")
-        console.print(f"   Recalled: {len(ctx.recalled_memorias)} memoria(s)")
-        console.print()
-
-
-@contextual_group.command(name="reset-preferences")
-@click.confirmation_option(prompt="This will reset all learned preferences. Continue?")
-def contextual_reset_preferences() -> None:
-    """Reset all learned user preferences.
-
-    Example: memo contextual reset-preferences
-    """
-    cfg = Config.from_env()
-
-    # Reset preferences file
-    prefs_file = cfg.state_dir / "user_preferences.json"
-    if prefs_file.is_file():
-        prefs_file.unlink()
-
-    # Reload will create fresh defaults
-    console.print("[green]Preferences reset successfully[/green]")
-
-
-# -- cross-reference commands -------------------------------------------------
-
-
-@cli.group(name="links")
-def links_group() -> None:
-    """Cross-reference and backlink system for memories."""
-    pass
-
-
-@links_group.command(name="backlinks")
-@click.argument("memoria_id")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def links_backlinks(memoria_id: str, as_json: bool) -> None:
-    """Show all memorias that reference this one.
-
-    Example: memo links backlinks abc123
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    backlinks = mem.crossref.get_backlinks(memoria_id)
-
-    if as_json:
-        click.echo(json.dumps([b.__dict__ for b in backlinks], indent=2))
-        return
-
-    if not backlinks:
-        console.print(f"[dim]No backlinks found for memoria {memoria_id[:8]}[/dim]")
-        return
-
-    table = Table(title=f"Backlinks to {memoria_id[:8]}")
-    table.add_column("Source ID", style="cyan")
-    table.add_column("Link Type", style="yellow")
-    table.add_column("Context", style="dim")
-
-    for bl in backlinks[:20]:
-        table.add_row(
-            bl.source_id[:8],
-            bl.link_type,
-            bl.context[:60] if bl.context else "—",
-        )
-
-    console.print(table)
-    if len(backlinks) > 20:
-        console.print(f"[dim]...and {len(backlinks) - 20} more[/dim]")
-
-
-@links_group.command(name="outlinks")
-@click.argument("memoria_id")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def links_outlinks(memoria_id: str, as_json: bool) -> None:
-    """Show all memorias that this one references.
-
-    Example: memo links outlinks abc123
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    outlinks = mem.crossref.get_outlinks(memoria_id)
-
-    if as_json:
-        click.echo(json.dumps([o.__dict__ for o in outlinks], indent=2))
-        return
-
-    if not outlinks:
-        console.print(f"[dim]No outlinks found for memoria {memoria_id[:8]}[/dim]")
-        return
-
-    console.print(f"[bold]Outlinks from {memoria_id[:8]}[/bold]")
-    console.print()
-
-    for ol in outlinks:
-        console.print(f"  [cyan]{ol.target}[/cyan]")
-
-
-@links_group.command(name="suggest")
-@click.argument("content")
-@click.option("--title", help="Title of the memoria being saved")
-@click.option("--tags", multiple=True, help="Tags of the memoria being saved")
-@click.option("--limit", type=int, default=5,
-              help="Max suggestions (default: 5)")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def links_suggest(content: str, title: str | None, tags: tuple[str, ...], limit: int, as_json: bool) -> None:
-    """Suggest links to existing memorias based on content.
-
-    Example: memo links suggest "MLX performance optimization" --title "MLX"
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    suggestions = mem.link_suggester.suggest_links(
-        content=content,
-        title=title or "",
-        tags=list(tags),
-        limit=limit,
-    )
-
-    if as_json:
-        click.echo(json.dumps([s.__dict__ for s in suggestions], indent=2))
-        return
-
-    if not suggestions:
-        console.print("[dim]No link suggestions found[/dim]")
-        return
-
-    table = Table(title="Link Suggestions")
-    table.add_column("ID", style="cyan")
-    table.add_column("Title", style="yellow")
-    table.add_column("Similarity", style="green")
-    table.add_column("Reason", style="dim")
-
-    for s in suggestions:
-        table.add_row(
-            s.memoria_id[:8],
-            s.title[:40],
-            f"{s.similarity:.3f}",
-            s.reason,
-        )
-
-    console.print(table)
-
-
-@links_group.command(name="format")
-@click.argument("memoria_id")
-@click.option("--title", help="Display title for the link")
-def links_format(memoria_id: str, title: str | None) -> None:
-    """Format a memoria ID as a wikilink.
-
-    Example: memo links format abc123 --title "My Memory"
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    wikilink = mem.link_suggester.format_wikilink(memoria_id, title)
-    click.echo(wikilink)
-
-
-@links_group.command(name="reindex")
-@click.confirmation_option(prompt="This will rebuild the entire cross-reference index. Continue?")
-def links_reindex() -> None:
-    """Rebuild the cross-reference index from all memorias.
-
-    Example: memo links reindex
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    # Delete existing index
-    cfg.crossref_db.unlink(missing_ok=True)
-
-    # Re-index all memorias
-    all_records = mem.list(limit=10000)
-    indexed = 0
-
-    for rec in all_records:
-        body = rec.body or ""
-        if body:
-            mem.crossref.index_wikilinks(rec.id, body)
-            indexed += 1
-
-    console.print(f"[green]Reindexed {indexed} memorias[/green]")
-
-
-# -- lifecycle management commands ---------------------------------------------
-
-
-@click.group(name="lifecycle")
-def lifecycle_group() -> None:
-    """Memory lifecycle management — archival, promotion, expiration."""
-    pass
-
-
-@lifecycle_group.command(name="report")
-@click.option("--limit", type=int, default=100,
-              help="Max memorias to analyze (default: 100)")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def lifecycle_report(limit: int, as_json: bool) -> None:
-    """Generate a lifecycle report on the corpus.
-
-    Shows statistics on archival candidates, promotion/demotion candidates,
-    expiration candidates, and access patterns.
-
-    Example: memo lifecycle report --limit 50
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    report = mem.lifecycle.get_lifecycle_report(limit=limit)
-
-    if as_json:
-        click.echo(json.dumps(report, indent=2))
-        return
-
-    console.print("[bold]Lifecycle Report[/bold]")
-    console.print()
-    console.print(f"Total memorias: {report['total']}")
-    console.print(f"Average access count: {report['avg_access_count']:.2f}")
-    console.print()
-    console.print(f"[yellow]Archive candidates:[/yellow] {report['archive_candidates']}")
-    console.print(f"[yellow]Promotion candidates:[/yellow] {report['promotion_candidates']}")
-    console.print(f"[yellow]Demotion candidates:[/yellow] {report['demotion_candidates']}")
-    console.print(f"[yellow]Expiration candidates:[/yellow] {report['expiration_candidates']}")
-    console.print(f"[yellow]Never accessed:[/yellow] {report['never_accessed']}")
-
-
-@lifecycle_group.command(name="apply")
-@click.option("--dry-run", is_flag=True,
-              help="Show what would happen without applying changes")
-@click.option("--limit", type=int, default=100,
-              help="Max memorias to process (default: 100)")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-@click.confirmation_option(prompt="This will archive/delete memorias based on lifecycle rules. Continue?")
-def lifecycle_apply(dry_run: bool, limit: int, as_json: bool) -> None:
-    """Apply lifecycle rules to the corpus.
-
-    Archives inactive memorias, expires temporary memories, and reports
-    promotion/demotion candidates. Use --dry-run first to preview.
-
-    Example: memo lifecycle apply --dry-run
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    if dry_run:
-        console.print("[yellow]Dry run mode - no changes will be applied[/yellow]")
-        console.print()
-
-    actions = mem.lifecycle.apply_lifecycle_rules(dry_run=dry_run, limit=limit)
-
-    if as_json:
-        click.echo(json.dumps(actions, indent=2))
-        return
-
-    console.print("[bold]Lifecycle Actions[/bold]")
-    console.print()
-    console.print(f"[green]Archived:[/green] {actions['archived']}")
-    console.print(f"[green]Promoted:[/green] {actions['promoted']}")
-    console.print(f"[yellow]Demoted:[/yellow] {actions['demoted']}")
-    console.print(f"[red]Expired:[/red] {actions['expired']}")
-    console.print(f"[red]Deleted:[/red] {actions['deleted']}")
-    console.print(f"[dim]Skipped:[/dim] {actions['skipped']}")
-
-
-@lifecycle_group.command(name="access-count")
-@click.argument("memoria_id")
-def lifecycle_access_count(memoria_id: str) -> None:
-    """Show access count for a specific memoria.
-
-    Example: memo lifecycle access-count abc123
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    count = mem.lifecycle.get_access_count(memoria_id)
-    console.print(f"Access count: {count}")
-
-
-@lifecycle_group.command(name="list-inactive")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def lifecycle_list_inactive(as_json: bool) -> None:
-    """List all archived/inactive memorias.
-
-    Example: memo lifecycle list-inactive
-    """
-    cfg = Config.from_env()
-    inactive_dir = cfg.memory_dir / "inactive"
-
-    if not inactive_dir.is_dir():
-        console.print("[dim]No inactive memorias found[/dim]")
-        return
-
-    files = list(inactive_dir.glob("*.md"))
-
-    if as_json:
-        inactive_data = []
-        for f in files:
-            inactive_data.append({
-                "id": f.stem,
-                "path": str(f),
-            })
-        click.echo(json.dumps(inactive_data, indent=2))
-        return
-
-    table = Table(title="Inactive Memorias")
-    table.add_column("ID", style="cyan")
-    table.add_column("Path", style="dim")
-
-    for f in files[:50]:
-        table.add_row(f.stem[:8], str(f.name))
-
-    console.print(table)
-    if len(files) > 50:
-        console.print(f"[dim]...and {len(files) - 50} more[/dim]")
-
-
-# -- proactive suggestions commands ---------------------------------------------
-
-
-@click.group(name="suggest")
-def suggest_group() -> None:
-    """Proactive memory suggestions from conversation analysis."""
-    pass
-
-
-@suggest_group.command(name="analyze")
-@click.argument("transcript_path", type=click.Path(exists=True))
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def suggest_analyze(transcript_path: str, as_json: bool) -> None:
-    """Analyze a transcript and suggest memories to save.
-
-    Example: memo suggest analyze /path/to/transcript.jsonl
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-
-    from memo.capture import _read_last_exchange
-
-    # For now, just use the last exchange as a sample
-    # In a full implementation, would analyze the full transcript
-    pair = _read_last_exchange(Path(transcript_path))
-    if pair is None:
-        console.print("[yellow]Could not read transcript[/yellow]")
-        return
-
-    user_text, assistant_text = pair
-    turns = [{"user": user_text, "assistant": assistant_text}]
-
-    suggestions = mem.proactive.analyze_conversation(turns, limit=3)
-
-    if as_json:
-        click.echo(json.dumps([s.__dict__ for s in suggestions], indent=2))
-        return
-
-    if not suggestions:
-        console.print("[dim]No suggestions found[/dim]")
-        return
-
-    console.print(f"[bold]Found {len(suggestions)} suggestions[/bold]")
-    console.print()
-
-    for i, s in enumerate(suggestions, 1):
-        console.print(f"[cyan]{i}. {s.title}[/cyan]")
-        console.print(f"   Type: {s.type}")
-        console.print(f"   Confidence: {s.confidence:.2f}")
-        console.print(f"   Tags: {', '.join(s.tags)}")
-        console.print(f"   Rationale: {s.rationale}")
-        console.print(f"   Snippet: {s.body_snippet[:100]}")
-        console.print()
-
-
-@suggest_group.command(name="feedback-stats")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def suggest_feedback_stats(as_json: bool) -> None:
-    """Show statistics on suggestion feedback (acceptance rate).
-
-    Example: memo suggest feedback-stats
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    stats = mem.proactive.get_feedback_stats()
-
-    if as_json:
-        click.echo(json.dumps(stats, indent=2))
-        return
-
-    console.print("[bold]Suggestion Feedback Stats[/bold]")
-    console.print()
-    console.print(f"Total suggestions: {stats['total']}")
-    console.print(f"Accepted: {stats['accepted']}")
-    console.print(f"Rejected: {stats['rejected']}")
-    console.print(f"Acceptance rate: {stats['acceptance_rate']:.2%}")
-
-
-@suggest_group.command(name="patterns")
-@click.argument("transcript_path", type=click.Path(exists=True))
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def suggest_patterns(transcript_path: str, as_json: bool) -> None:
-    """Detect patterns in a transcript (recurring themes, decisions, etc.).
-
-    Example: memo suggest patterns /path/to/transcript.jsonl
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-
-    patterns = mem.proactive.detect_patterns(Path(transcript_path))
-
-    if as_json:
-        click.echo(json.dumps(patterns, indent=2))
-        return
-
-    console.print("[bold]Conversation Patterns[/bold]")
-    console.print()
-    console.print(f"Total turns: {patterns['total_turns']}")
-    console.print(f"Decision points: {patterns['decision_points']}")
-    console.print(f"Technical discoveries: {patterns['technical_discoveries']}")
-    console.print(f"Recurring themes: {', '.join(patterns['recurring_themes'])}")
-
-
-# -- versioning commands ------------------------------------------------------
-
-
-@cli.group(name="version")
-def version_group() -> None:
-    """Memory versioning — track changes, visualize diffs, rollback."""
-    pass
-
-
-@version_group.command(name="history")
-@click.argument("memoria_id")
-@click.option("--limit", type=int, default=10,
-              help="Max versions to show (default: 10)")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def version_history(memoria_id: str, limit: int, as_json: bool) -> None:
-    """Show version history for a memoria.
-
-    Example: memo version history abc123
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    versions = mem.versioning.get_version_history(memoria_id, limit=limit)
-
-    if as_json:
-        click.echo(json.dumps([v.__dict__ for v in versions], indent=2))
-        return
-
-    if not versions:
-        console.print(f"[dim]No version history for memoria {memoria_id[:8]}[/dim]")
-        return
-
-    console.print(f"[bold]Version History for {memoria_id[:8]}[/bold]")
-    console.print()
-
-    table = Table()
-    table.add_column("Version ID", style="cyan")
-    table.add_column("Timestamp", style="dim")
-    table.add_column("Title", style="yellow")
-    table.add_column("Type", style="green")
-    table.add_column("Reason", style="magenta")
-
-    for v in versions[:20]:
-        table.add_row(
-            str(v.version_id),
-            v.timestamp[:19],
-            v.title[:40],
-            v.type,
-            v.reason or "—",
-        )
-
-    console.print(table)
-    if len(versions) > 20:
-        console.print(f"[dim]...and {len(versions) - 20} more[/dim]")
-
-
-@version_group.command(name="diff")
-@click.argument("memoria_id")
-@click.option("--version-a", type=int, help="First version ID (default: latest)")
-@click.option("--version-b", type=int, help="Second version ID (default: latest-1)")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def version_diff(memoria_id: str, version_a: int | None, version_b: int | None, as_json: bool) -> None:
-    """Show diff between two versions of a memoria.
-
-    Example: memo version diff abc123 --version-a 1 --version-b 2
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    diff = mem.versioning.diff_versions(memoria_id, version_a, version_b)
-
-    if as_json:
-        click.echo(json.dumps(diff.__dict__ if diff else None, indent=2))
-        return
-
-    if diff is None:
-        console.print("[yellow]Could not generate diff[/yellow]")
-        return
-
-    console.print(f"[bold]Diff for {memoria_id[:8]}[/bold]")
-    console.print(f"[dim]v{diff.version_a} → v{diff.version_b}[/dim]")
-    console.print()
-    console.print(diff.unified_diff)
-
-
-@version_group.command(name="rollback")
-@click.argument("memoria_id")
-@click.argument("version_id", type=int)
-@click.option("--reason", help="Reason for the rollback")
-@click.confirmation_option(prompt="This will restore the memoria to the specified version. Continue?")
-def version_rollback(memoria_id: str, version_id: int, reason: str | None) -> None:
-    """Rollback a memoria to a previous version.
-
-    Example: memo version rollback abc123 1 --reason "Mistake in update"
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    success = mem.versioning.rollback_to_version(memoria_id, version_id, reason)
-
-    if success:
-        console.print(f"[green]Rolled back {memoria_id[:8]} to version {version_id}[/green]")
-    else:
-        console.print("[red]Failed to rollback[/red]")
-
-
-# -- query composition commands ------------------------------------------------
-
-
-@cli.group(name="query")
-def query_group() -> None:
-    """Query composition and saved queries."""
-    pass
-
-
-@query_group.command(name="save")
-@click.argument("name")
-@click.argument("query_text")
-@click.option("--type", "type_filter", help="Filter by memoria type")
-@click.option("--tags", "tags_filter", multiple=True, help="Filter by tags")
-@click.option("--date-from", help="Start date (ISO format)")
-@click.option("--date-to", help="End date (ISO format)")
-@click.option("--mode", "search_mode", type=click.Choice(["vec", "bm25", "hybrid"]), default="hybrid",
-              help="Search mode (default: hybrid)")
-@click.option("--limit", type=int, default=10, help="Result limit")
-@click.option("--description", help="Query description")
-@click.option("--execute", is_flag=True, help="Execute the query after saving")
-def query_save(
-    name: str,
-    query_text: str,
-    type_filter: str | None,
-    tags_filter: tuple[str, ...],
-    date_from: str | None,
-    date_to: str | None,
-    search_mode: str,
-    limit: int,
-    description: str | None,
-    execute: bool,
-) -> None:
-    """Save a query for reuse.
-
-    Example: memo query save "MLX decisions" "MLX" --type decision --execute
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    if execute:
-        result = mem.query_composer.compose_and_save(
-            name=name,
-            query_text=query_text,
-            type_filter=type_filter,
-            tags_filter=list(tags_filter),
-            date_from=date_from,
-            date_to=date_to,
-            search_mode=search_mode,
-            limit=limit,
-            description=description,
-        )
-        console.print(f"[green]Saved and executed query '{name}'[/green]")
-        console.print(f"Results: {result.count}")
-    else:
-        mem.query_composer.query_store.save_query(
-            name=name,
-            query_text=query_text,
-            type_filter=type_filter,
-            tags_filter=list(tags_filter),
-            date_from=date_from,
-            date_to=date_to,
-            search_mode=search_mode,
-            limit=limit,
-            description=description,
-        )
-        console.print(f"[green]Saved query '{name}'[/green]")
-
-
-@query_group.command(name="list")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def query_list(as_json: bool) -> None:
-    """List all saved queries.
-
-    Example: memo query list
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    queries = mem.query_composer.query_store.list_queries()
-
-    if as_json:
-        click.echo(json.dumps([q.__dict__ for q in queries], indent=2))
-        return
-
-    if not queries:
-        console.print("[dim]No saved queries[/dim]")
-        return
-
-    table = Table(title="Saved Queries")
-    table.add_column("Name", style="cyan")
-    table.add_column("Query Text", style="yellow")
-    table.add_column("Type Filter", style="green")
-    table.add_column("Mode", style="magenta")
-    table.add_column("Description", style="dim")
-
-    for q in queries[:20]:
-        table.add_row(
-            q.name,
-            q.query_text[:40],
-            q.type_filter or "—",
-            q.search_mode,
-            q.description or "—",
-        )
-
-    console.print(table)
-    if len(queries) > 20:
-        console.print(f"[dim]...and {len(queries) - 20} more[/dim]")
-
-
-@query_group.command(name="run")
-@click.argument("name")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def query_run(name: str, as_json: bool) -> None:
-    """Execute a saved query.
-
-    Example: memo query run "MLX decisions"
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    query = mem.query_composer.query_store.get_query(name)
-    if not query:
-        console.print(f"[yellow]Query '{name}' not found[/yellow]")
-        return
-
-    result = mem.query_composer.execute_query(query)
-
-    if as_json:
-        # Convert results to dict format
-        results_dict = [r.__dict__ for r in result.results]
-        click.echo(json.dumps({
-            "query_name": result.query_name,
-            "count": result.count,
-            "executed_at": result.executed_at,
-            "results": results_dict,
-        }, indent=2))
-        return
-
-    console.print(f"[bold]Query: {name}[/bold]")
-    console.print(f"Results: {result.count}")
-    console.print()
-
-    for r in result.results[:10]:
-        console.print(f"  [cyan]{r.id[:8]}[/cyan] {r.title}")
-        console.print(f"    {r.body[:100]}")
-
-    if len(result.results) > 10:
-        console.print(f"  [dim]...and {len(result.results) - 10} more[/dim]")
-
-
-@query_group.command(name="delete")
-@click.argument("name")
-@click.confirmation_option(prompt="Delete this saved query?")
-def query_delete(name: str) -> None:
-    """Delete a saved query.
-
-    Example: memo query delete "MLX decisions"
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    success = mem.query_composer.query_store.delete_query(name)
-
-    if success:
-        console.print(f"[green]Deleted query '{name}'[/green]")
-    else:
-        console.print(f"[yellow]Query '{name}' not found[/yellow]")
-
-
-# -- federation commands -------------------------------------------------------
-
-
-@click.group(name="federation")
-def federation_group() -> None:
-    """Multi-vault federation — search across multiple vaults."""
-    pass
-
-
-@federation_group.command(name="add-vault")
-@click.argument("name")
-@click.argument("path")
-@click.option("--weight", type=float, default=1.0, help="Vault weight for ranking")
-def federation_add_vault(name: str, path: str, weight: float) -> None:
-    """Add a vault to the federation.
-
-    Example: memo federation add-vault work-vault /path/to/work/memo --weight 1.5
-    """
-    cfg = Config.from_env()
-    from memo.federation import FederationConfig
-
-    config = FederationConfig(cfg.state_dir / "federation.json")
-    config.add_vault(name, path, weight)
-
-    console.print(f"[green]Added vault '{name}'[/green]")
-
-
-@federation_group.command(name="list-vaults")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def federation_list_vaults(as_json: bool) -> None:
-    """List all configured vaults.
-
-    Example: memo federation list-vaults
-    """
-    cfg = Config.from_env()
-    from memo.federation import FederationConfig
-
-    config = FederationConfig(cfg.state_dir / "federation.json")
-    vaults = config.list_vaults()
-
-    if as_json:
-        click.echo(json.dumps([v.__dict__ for v in vaults], indent=2))
-        return
-
-    if not vaults:
-        console.print("[dim]No vaults configured[/dim]")
-        return
-
-    table = Table(title="Federated Vaults")
-    table.add_column("Name", style="cyan")
-    table.add_column("Path", style="yellow")
-    table.add_column("Weight", style="green")
-    table.add_column("Enabled", style="magenta")
-
-    for v in vaults:
-        table.add_row(
-            v.name,
-            v.path,
-            str(v.weight),
-            "Yes" if v.enabled else "No",
-        )
-
-    console.print(table)
-
-
-@federation_group.command(name="remove-vault")
-@click.argument("name")
-@click.confirmation_option(prompt="Remove this vault from federation?")
-def federation_remove_vault(name: str) -> None:
-    """Remove a vault from the federation.
-
-    Example: memo federation remove-vault work-vault
-    """
-    cfg = Config.from_env()
-    from memo.federation import FederationConfig
-
-    config = FederationConfig(cfg.state_dir / "federation.json")
-    success = config.remove_vault(name)
-
-    if success:
-        console.print(f"[green]Removed vault '{name}'[/green]")
-    else:
-        console.print(f"[yellow]Vault '{name}' not found[/yellow]")
-
-
-@federation_group.command(name="search")
-@click.argument("query")
-@click.option("--limit", type=int, default=10, help="Result limit")
-@click.option("--mode", type=click.Choice(["vec", "bm25", "hybrid"]), default="hybrid",
-              help="Search mode (default: hybrid)")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def federation_search(query: str, limit: int, mode: str, as_json: bool) -> None:
-    """Search across all federated vaults.
-
-    Example: memo federation search "MLX" --limit 20
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    results = mem.federation.search(query, limit=limit, mode=mode)
-
-    if as_json:
-        click.echo(json.dumps([r.__dict__ for r in results], indent=2))
-        return
-
-    if not results:
-        console.print("[dim]No results found[/dim]")
-        return
-
-    table = Table(title=f"Federated Search Results for '{query}'")
-    table.add_column("ID", style="cyan")
-    table.add_column("Vault", style="yellow")
-    table.add_column("Title", style="green")
-    table.add_column("Score", style="magenta")
-
-    for r in results[:20]:
-        table.add_row(
-            r.memoria_id[:8],
-            r.vault_name,
-            r.title[:40],
-            f"{r.score:.3f}",
-        )
-
-    console.print(table)
-    if len(results) > 20:
-        console.print(f"[dim]...and {len(results) - 20} more[/dim]")
-
-
-# -- sync & backup commands ------------------------------------------------------
-
-
-@cli.group(name="backup", invoke_without_command=True)
-@click.option("--out", "out_path", default=None, type=click.Path(),
-              help="Output portable zip path. Default: ./memo-backup-<YYYYMMDD-HHMMSS>.zip")
-@click.pass_context
-def backup_group(ctx: click.Context, out_path: str | None) -> None:
-    """Create a portable backup, or manage named backups with subcommands."""
-    if ctx.invoked_subcommand is None:
-        _portable_backup(out_path)
-    pass
-
-
-@backup_group.command(name="create")
-@click.option("--compress/--no-compress", default=True, help="Compress backup")
-@click.option("--name", help="Backup name")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def backup_create(compress: bool, name: str | None, as_json: bool) -> None:
-    """Create a backup of the entire vault.
-
-    Example: memo backup create --name "pre-migration"
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    metadata = mem.backup.create_backup(compress=compress, name=name)
-
-    if as_json:
-        click.echo(json.dumps(metadata.__dict__, indent=2))
-        return
-
-    console.print("[bold]Backup Created[/bold]")
-    console.print()
-    console.print(f"Timestamp: {metadata.timestamp}")
-    console.print(f"Memorias: {metadata.memoria_count}")
-    console.print(f"Checksum: {metadata.checksum[:16]}...")
-    console.print(f"Size: {metadata.compressed_size:,} bytes (compressed)")
-    console.print(f"Original: {metadata.original_size:,} bytes")
-
-
-@backup_group.command(name="list")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def backup_list(as_json: bool) -> None:
-    """List all available backups.
-
-    Example: memo backup list
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    backups = mem.backup.list_backups()
-
-    if as_json:
-        click.echo(json.dumps([b.__dict__ for b in backups], indent=2))
-        return
-
-    if not backups:
-        console.print("[dim]No backups found[/dim]")
-        return
-
-    table = Table(title="Available Backups")
-    table.add_column("Name", style="cyan")
-    table.add_column("Timestamp", style="yellow")
-    table.add_column("Size", style="green")
-
-    for b in backups[:20]:
-        table.add_row(
-            b.timestamp[:19],
-            b.timestamp[:19],
-            f"{b.compressed_size:,} bytes",
-        )
-
-    console.print(table)
-    if len(backups) > 20:
-        console.print(f"[dim]...and {len(backups) - 20} more[/dim]")
-
-
-@backup_group.command(name="restore")
-@click.argument("backup_name")
-@click.option("--no-memorias", "skip_memorias", is_flag=True, help="Skip memoria files")
-@click.option("--no-dbs", "skip_dbs", is_flag=True, help="Skip databases")
-@click.confirmation_option(prompt="This will restore from backup. Current data may be overwritten. Continue?")
-def backup_restore(backup_name: str, skip_memorias: bool, skip_dbs: bool) -> None:
-    """Restore from a backup.
-
-    Example: memo backup restore backup_2026-01-01-12-00-00
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    success = mem.backup.restore_backup(
-        backup_name,
-        restore_memorias=not skip_memorias,
-        restore_dbs=not skip_dbs,
-    )
-
-    if success:
-        console.print(f"[green]Restored from '{backup_name}'[/green]")
-    else:
-        console.print("[red]Failed to restore[/red]")
-
-
-@cli.group(name="sync")
-def sync_group() -> None:
-    """Multi-vault sync — sync between vaults."""
-    pass
-
-
-@sync_group.command(name="diff")
-@click.option("--remote", help="Path to remote vault")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def sync_diff(remote: str | None, as_json: bool) -> None:
-    """Compute diff between local and remote vaults.
-
-    Example: memo sync diff --remote /path/to/remote/memo
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-
-    from memo.sync import SyncManager
-    remote_path = Path(remote) if remote else None
-
-    sync_mgr = SyncManager(mem, remote_path=remote_path)
-    diff = sync_mgr.compute_diff()
-
-    if as_json:
-        click.echo(json.dumps(diff.__dict__, indent=2))
-        return
-
-    console.print("[bold]Sync Diff[/bold]")
-    console.print()
-    console.print(f"New: {len(diff.new)}")
-    console.print(f"Modified: {len(diff.modified)}")
-    console.print(f"Deleted: {len(diff.deleted)}")
-    console.print(f"Conflicts: {len(diff.conflicts)}")
-
-
-@sync_group.command(name="push")
-@click.option("--remote", help="Path to remote vault")
-def sync_push(remote: str | None) -> None:
-    """Push local changes to remote vault.
-
-    Example: memo sync push --remote /path/to/remote/memo
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-
-    from memo.sync import SyncManager
-    remote_path = Path(remote) if remote else None
-
-    sync_mgr = SyncManager(mem, remote_path=remote_path)
-    diff = sync_mgr.sync(direction="push")
-
-    console.print("[bold]Push Sync[/bold]")
-    console.print(f"Modified: {len(diff.modified)}")
-    console.print(f"Deleted: {len(diff.deleted)}")
-
-
-@sync_group.command(name="pull")
-@click.option("--remote", help="Path to remote vault")
-def sync_pull(remote: str | None) -> None:
-    """Pull remote changes to local vault.
-
-    Example: memo sync pull --remote /path/to/remote/memo
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-
-    from memo.sync import SyncManager
-    remote_path = Path(remote) if remote else None
-
-    sync_mgr = SyncManager(mem, remote_path=remote_path)
-    diff = sync_mgr.sync(direction="pull")
-
-    console.print("[bold]Pull Sync[/bold]")
-    console.print(f"New: {len(diff.new)}")
-    console.print(f"Modified: {len(diff.modified)}")
-
-
-@sync_group.command(name="both")
-@click.option("--remote", help="Path to remote vault")
-def sync_both(remote: str | None) -> None:
-    """Sync both directions (bidirectional).
-
-    Example: memo sync both --remote /path/to/remote/memo
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-
-    from memo.sync import SyncManager
-    remote_path = Path(remote) if remote else None
-
-    sync_mgr = SyncManager(mem, remote_path=remote_path)
-    diff = sync_mgr.sync(direction="both")
-
-    console.print("[bold]Bidirectional Sync[/bold]")
-    console.print(f"New: {len(diff.new)}")
-    console.print(f"Modified: {len(diff.modified)}")
-    console.print(f"Deleted: {len(diff.deleted)}")
-    console.print(f"Conflicts: {len(diff.conflicts)}")
-
-
-# -- encryption commands ----------------------------------------------------------
-
-
-@cli.group(name="encrypt")
-def encrypt_group() -> None:
-    """Memory encryption — encrypt sensitive memorias."""
-    pass
-
-
-@encrypt_group.command(name="unlock")
-@click.argument("password")
-def encrypt_unlock(password: str) -> None:
-    """Unlock the vault with password.
-
-    Derives master key from password and stores in memory.
-
-    Example: memo encrypt unlock mypassword
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    success = mem.encryption.unlock(password)
-
-    if success:
-        console.print("[green]Vault unlocked[/green]")
-    else:
-        console.print("[red]Failed to unlock vault[/red]")
-
-
-@encrypt_group.command(name="lock")
-def encrypt_lock() -> None:
-    """Lock the vault (clear master key from memory).
-
-    Example: memo encrypt lock
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    mem.encryption.lock()
-    console.print("[green]Vault locked[/green]")
-
-
-@encrypt_group.command(name="status")
-def encrypt_status() -> None:
-    """Check if vault is unlocked.
-
-    Example: memo encrypt status
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    if mem.encryption.is_unlocked():
-        console.print("[green]Vault is unlocked[/green]")
-    else:
-        console.print("[yellow]Vault is locked[/yellow]")
-
-
-# -- sharing commands -----------------------------------------------------------
-
-
-@cli.group(name="share")
-def share_group() -> None:
-    """Memory sharing — share memorias with others."""
-    pass
-
-
-@share_group.command(name="with-user")
-@click.argument("memoria_id")
-@click.argument("shared_with")
-@click.option("--permission", type=click.Choice(["read", "comment", "edit", "admin"]), default="read",
-              help="Permission level")
-@click.option("--expires-days", type=int, help="Days until expiration")
-def share_with_user(memoria_id: str, shared_with: str, permission: str, expires_days: int | None) -> None:
-    """Share a memoria with a user.
-
-    Example: memo share with-user abc123 user@example.com --permission comment
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    share = mem.sharing.share_with_user(
-        memoria_id=memoria_id,
-        shared_with=shared_with,
-        permission=permission,
-        expires_days=expires_days,
-    )
-
-    console.print(f"[green]Shared {memoria_id[:8]} with {shared_with}[/green]")
-    console.print(f"Permission: {permission}")
-    if share.expires_at:
-        console.print(f"Expires: {share.expires_at}")
-
-
-@share_group.command(name="unshare")
-@click.argument("memoria_id")
-@click.argument("shared_with")
-def share_unshare(memoria_id: str, shared_with: str) -> None:
-    """Unshare a memoria from a user.
-
-    Example: memo share unshare abc123 user@example.com
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    success = mem.sharing.unshare_with_user(memoria_id, shared_with)
-
-    if success:
-        console.print(f"[green]Unshared {memoria_id[:8]} from {shared_with}[/green]")
-    else:
-        console.print("[yellow]Share not found[/yellow]")
-
-
-@share_group.command(name="create-link")
-@click.argument("memoria_id")
-@click.option("--permission", type=click.Choice(["read", "comment", "edit"]), default="read",
-              help="Permission level")
-@click.option("--expires-hours", type=int, default=24, help="Hours until expiration")
-@click.option("--password", help="Optional password protection")
-def share_create_link(memoria_id: str, permission: str, expires_hours: int, password: str | None) -> None:
-    """Create a temporary sharing link.
-
-    Example: memo share create-link abc123 --permission read --expires-hours 48
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    link = mem.sharing.create_link(
-        memoria_id=memoria_id,
-        permission=permission,
-        expires_hours=expires_hours,
-        password=password,
-    )
-
-    console.print("[green]Share link created[/green]")
-    console.print(f"Link: {link}")
-    console.print(f"Expires in {expires_hours} hours")
-
-
-@share_group.command(name="list")
-@click.argument("memoria_id")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def share_list(memoria_id: str, as_json: bool) -> None:
-    """List all shares for a memoria.
-
-    Example: memo share list abc123
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    shares = mem.sharing.share_store.get_shares(memoria_id)
-
-    if as_json:
-        click.echo(json.dumps([s.__dict__ for s in shares], indent=2))
-        return
-
-    if not shares:
-        console.print("[dim]No shares found[/dim]")
-        return
-
-    table = Table(title=f"Shares for {memoria_id[:8]}")
-    table.add_column("Shared With", style="cyan")
-    table.add_column("Permission", style="yellow")
-    table.add_column("Shared At", style="green")
-    table.add_column("Expires", style="magenta")
-
-    for s in shares[:20]:
-        table.add_row(
-            s.shared_with,
-            s.permission,
-            s.shared_at[:19],
-            s.expires_at[:19] if s.expires_at else "Never",
-        )
-
-    console.print(table)
-    if len(shares) > 20:
-        console.print(f"[dim]...and {len(shares) - 20} more[/dim]")
-
-
-@share_group.command(name="comment")
-@click.argument("memoria_id")
-@click.argument("content")
-@click.option("--author", default="user", help="Comment author")
-@click.option("--parent", help="Parent comment ID for replies")
-def share_comment(memoria_id: str, content: str, author: str, parent: str | None) -> None:
-    """Add a comment to a memoria.
-
-    Example: memo share comment abc123 "This is a comment" --author "John Doe"
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    comment = mem.sharing.add_comment(
-        memoria_id=memoria_id,
-        author=author,
-        content=content,
-        parent_id=parent,
-    )
-
-    console.print("[green]Comment added[/green]")
-    console.print(f"Author: {comment.author}")
-    console.print(f"Content: {comment.content}")
-
-
-@share_group.command(name="comments")
-@click.argument("memoria_id")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def share_comments(memoria_id: str, as_json: bool) -> None:
-    """List all comments for a memoria.
-
-    Example: memo share comments abc123
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    comments = mem.sharing.get_comments(memoria_id)
-
-    if as_json:
-        click.echo(json.dumps([c.__dict__ for c in comments], indent=2))
-        return
-
-    if not comments:
-        console.print("[dim]No comments found[/dim]")
-        return
-
-    table = Table(title=f"Comments for {memoria_id[:8]}")
-    table.add_column("Author", style="cyan")
-    table.add_column("Content", style="yellow")
-    table.add_column("Created", style="green")
-
-    for c in comments[:20]:
-        table.add_row(
-            c.author,
-            c.content[:50],
-            c.created_at[:19],
-        )
-
-    console.print(table)
-    if len(comments) > 20:
-        console.print(f"[dim]...and {len(comments) - 20} more[/dim]")
-
-
-# -- analytics commands ----------------------------------------------------------
-
-
-@cli.group(name="analytics")
-def analytics_group() -> None:
-    """Memory analytics dashboard — metrics and visualizations."""
-    pass
-
-
-@analytics_group.command(name="summary")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def analytics_summary(as_json: bool) -> None:
-    """Show analytics summary.
-
-    Example: memo analytics summary
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    metrics = mem.analytics.compute_corpus_metrics()
-
-    if as_json:
-        click.echo(json.dumps(metrics.__dict__, indent=2))
-        return
-
-    console.print("[bold]Memory Analytics Summary[/bold]")
-    console.print()
-    console.print(f"Total Memorias: {metrics.total_memorias}")
-    console.print(f"Total Entities: {metrics.total_entities}")
-    console.print(f"Growth Rate: {metrics.growth_rate:.2f} memorias/day")
-    console.print(f"Average Access Count: {metrics.average_access_count:.2f}")
-    console.print()
-    console.print("[bold]Type Distribution[/bold]")
-    for t, c in metrics.type_distribution.items():
-        console.print(f"  {t}: {c}")
-    console.print()
-    console.print("[bold]Top 10 Tags[/bold]")
-    for i, (t, c) in enumerate(list(metrics.tag_frequency.items())[:10], 1):
-        console.print(f"  {i}. {t}: {c}")
-
-
-@analytics_group.command(name="growth")
-@click.option("--days", type=int, default=30, help="Days to analyze")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def analytics_growth(days: int, as_json: bool) -> None:
-    """Show growth data over time.
-
-    Example: memo analytics growth --days 30
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    growth = mem.analytics.compute_growth_data(days=days)
-
-    if as_json:
-        click.echo(json.dumps(growth.__dict__, indent=2))
-        return
-
-    console.print(f"[bold]Growth (Last {days} Days)[/bold]")
-    console.print()
-
-    table = Table()
-    table.add_column("Date", style="cyan")
-    table.add_column("Count", style="green")
-
-    for d, c in zip(growth.dates, growth.counts, strict=False):
-        table.add_row(d, str(c))
-
-    console.print(table)
-
-
-@analytics_group.command(name="export-json")
-@click.argument("output_path", type=click.Path())
-def analytics_export_json(output_path: str) -> None:
-    """Export analytics to JSON.
-
-    Example: memo analytics export-json /path/to/analytics.json
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-    mem.analytics.export_metrics_json(Path(output_path))
-
-    console.print(f"[green]Exported analytics to {output_path}[/green]")
-
-
-@analytics_group.command(name="export-csv")
-@click.argument("output_path", type=click.Path())
-def analytics_export_csv(output_path: str) -> None:
-    """Export analytics to CSV.
-
-    Example: memo analytics export-csv /path/to/analytics.csv
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-    mem.analytics.export_metrics_csv(Path(output_path))
-
-    console.print(f"[green]Exported analytics to {output_path}[/green]")
-
-
-@analytics_group.command(name="dashboard-html")
-@click.argument("output_path", type=click.Path())
-def analytics_dashboard_html(output_path: str) -> None:
-    """Generate HTML dashboard.
-
-    Example: memo analytics dashboard-html /path/to/dashboard.html
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-    mem.dashboard.generate_html_dashboard(Path(output_path))
-
-    console.print(f"[green]Generated HTML dashboard at {output_path}[/green]")
-
-
-# -- import/export commands ------------------------------------------------------
-
-
-@cli.group(name="import")
-def import_group() -> None:
-    """Import memorias from other formats."""
-    pass
-
-
-@import_group.command(name="json")
-@click.argument("input_path", type=click.Path())
-@click.option("--format", help="Format (json, csv, markdown_bundle)")
-def import_json(input_path: str, format: str | None) -> None:
-    """Import from JSON file.
-
-    Example: memo import json /path/to/export.json
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-    result = mem.import_export.import_from(Path(input_path), format or "json")
-
-    console.print("[green]Import complete[/green]")
-    console.print(f"Imported: {result.imported_count}")
-    console.print(f"Skipped: {result.skipped_count}")
-
-    if result.errors:
-        console.print(f"[yellow]Errors: {len(result.errors)}[/yellow]")
-
-
-@import_group.command(name="csv")
-@click.argument("input_path", type=click.Path())
-def import_csv(input_path: str) -> None:
-    """Import from CSV file.
-
-    Example: memo import csv /path/to/export.csv
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-    result = mem.import_export.import_from(Path(input_path), "csv")
-
-    console.print("[green]Import complete[/green]")
-    console.print(f"Imported: {result.imported_count}")
-    console.print(f"Skipped: {result.skipped_count}")
-
-    if result.errors:
-        console.print(f"[yellow]Errors: {len(result.errors)}[/yellow]")
-
-
-@import_group.command(name="markdown-bundle")
-@click.argument("input_path", type=click.Path())
-def import_markdown_bundle(input_path: str) -> None:
-    """Import from Markdown bundle (zip).
-
-    Example: memo import markdown-bundle /path/to/export.zip
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-    result = mem.import_export.import_from(Path(input_path), "markdown_bundle")
-
-    console.print("[green]Import complete[/green]")
-    console.print(f"Imported: {result.imported_count}")
-    console.print(f"Skipped: {result.skipped_count}")
-
-    if result.errors:
-        console.print(f"[yellow]Errors: {len(result.errors)}[/yellow]")
-
-
-@cli.group(name="export")
-def export_group() -> None:
-    """Export memorias to other formats."""
-    pass
-
-
-@export_group.command(name="json")
-@click.argument("output_path", type=click.Path())
-def export_json(output_path: str) -> None:
-    """Export to JSON file.
-
-    Example: memo export json /path/to/export.json
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-    result = mem.import_export.export_to(Path(output_path), "json")
-
-    console.print("[green]Export complete[/green]")
-    console.print(f"Exported: {result.exported_count}")
-    console.print(f"Output: {result.output_path}")
-
-
-@export_group.command(name="csv")
-@click.argument("output_path", type=click.Path())
-def export_csv(output_path: str) -> None:
-    """Export to CSV file.
-
-    Example: memo export csv /path/to/export.csv
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-    result = mem.import_export.export_to(Path(output_path), "csv")
-
-    console.print("[green]Export complete[/green]")
-    console.print(f"Exported: {result.exported_count}")
-    console.print(f"Output: {result.output_path}")
-
-
-@export_group.command(name="markdown-bundle")
-@click.argument("output_path", type=click.Path())
-def export_markdown_bundle(output_path: str) -> None:
-    """Export to Markdown bundle (zip).
-
-    Example: memo export markdown-bundle /path/to/export.zip
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-    result = mem.import_export.export_to(Path(output_path), "markdown_bundle")
-
-    console.print("[green]Export complete[/green]")
-    console.print(f"Exported: {result.exported_count}")
-    console.print(f"Output: {result.output_path}")
-
-
-# -- multi-modal commands (gamechanger #17) ---------------------------------------
-
-
-@cli.group(name="multimodal")
-def multimodal_group() -> None:
-    """Memoria Multi-Modal con Embeddings Universales."""
-    pass
-
-
-@multimodal_group.command(name="add-image")
-@click.argument("image_path", type=click.Path(exists=True))
-@click.option("--memoria-id", help="ID de memoria asociada")
-def multimodal_add_image(image_path: str, memoria_id: str | None) -> None:
-    """Agrega imagen al corpus multi-modal.
-
-    Example: memo multimodal add-image /path/to/image.png --memoria-id abc123
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-    content = mem.multimodal.add_image(Path(image_path), memoria_id)
-
-    console.print("[green]Image added[/green]")
-    console.print(f"Content ID: {content.id}")
-    console.print(f"Modality: {content.modality}")
-
-
-@multimodal_group.command(name="add-audio")
-@click.argument("audio_path", type=click.Path(exists=True))
-@click.option("--memoria-id", help="ID de memoria asociada")
-def multimodal_add_audio(audio_path: str, memoria_id: str | None) -> None:
-    """Agrega audio al corpus multi-modal.
-
-    Example: memo multimodal add-audio /path/to/audio.mp3 --memoria-id abc123
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    from pathlib import Path
-    content = mem.multimodal.add_audio(Path(audio_path), memoria_id)
-
-    console.print("[green]Audio added[/green]")
-    console.print(f"Content ID: {content.id}")
-    console.print(f"Modality: {content.modality}")
-
-
-@multimodal_group.command(name="search-images")
-@click.argument("query")
-@click.option("--limit", type=int, default=10, help="Máximo de resultados")
-def multimodal_search_images(query: str, limit: int) -> None:
-    """Busca con texto, encuentra imágenes.
-
-    Example: memo multimodal search-images "architecture diagram"
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    results = mem.multimodal.search.search_text_find_images(query, limit=limit)
-
-    console.print(f"[bold]Results: {len(results)} images[/bold]")
-    for r in results:
-        console.print(f"  {r.content_id[:8]} - similarity: {r.similarity:.2f}")
-
-
-@multimodal_group.command(name="search-audio")
-@click.argument("query")
-@click.option("--limit", type=int, default=10, help="Máximo de resultados")
-def multimodal_search_audio(query: str, limit: int) -> None:
-    """Busca con texto, encuentra audio.
-
-    Example: memo multimodal search-audio "meeting notes"
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    results = mem.multimodal.search.search_text_find_audio(query, limit=limit)
-
-    console.print(f"[bold]Results: {len(results)} audio[/bold]")
-    for r in results:
-        console.print(f"  {r.content_id[:8]} - similarity: {r.similarity:.2f}")
-
-
-@multimodal_group.command(name="search-all")
-@click.argument("query")
-@click.option("--limit", type=int, default=10, help="Máximo de resultados por modalidad")
-def multimodal_search_all(query: str, limit: int) -> None:
-    """Busca en todas las modalidades.
-
-    Example: memo multimodal search-all "project documentation"
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    results = mem.multimodal.search.search_all_modalities(query, limit=limit)
-
-    console.print("[bold]Results across modalities[/bold]")
-    for modality, mod_results in results.items():
-        console.print(f"\n[cyan]{modality}:[/cyan] {len(mod_results)} results")
-    for r in mod_results[:5]:
-        console.print(f"  {r.content_id[:8]} - similarity: {r.similarity:.2f}")
-
-
-# -- collaborative commands (gamechanger #18) -----------------------------------
-
-
-@cli.group(name="collaborative")
-def collaborative_group() -> None:
-    """Memoria Social Colaborativa con Grafo de Conocimiento Compartido."""
-    pass
-
-
-@collaborative_group.command(name="share-connection")
-@click.argument("user-id")
-@click.argument("entity-a")
-@click.argument("entity-b")
-@click.argument("relationship")
-@click.option("--confidence", type=float, default=0.7, help="Confidence score")
-def collaborative_share_connection(user_id: str, entity_a: str, entity_b: str, relationship: str, confidence: float) -> None:
-    """Comparte una conexión descubierta con la comunidad.
-
-    Example: memo collaborative share-connection user123 MLX Apple "optimized for"
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    conn = mem.collaborative.share_connection(
-        user_id=user_id,
-        entity_a=entity_a,
-        entity_b=entity_b,
-        relationship=relationship,
-        confidence=confidence,
-    )
-
-    console.print("[green]Connection shared[/green]")
-    console.print(f"Connection ID: {conn.connection_id}")
-    console.print(f"From: {conn.from_user}")
-    console.print(f"{entity_a} --{relationship}--> {entity_b}")
-
-
-@collaborative_group.command(name="connections")
-@click.argument("entity")
-def collaborative_connections(entity: str) -> None:
-    """Ver conexiones compartidas para una entidad.
-
-    Example: memo collaborative connections MLX
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    connections = mem.collaborative.get_shared_connections(entity)
-
-    console.print(f"[bold]Shared connections for {entity}[/bold]")
-    for c in connections:
-        console.print(f"  {c.entity_a} --{c.relationship}--> {c.entity_b}")
-        console.print(f"    From: {c.from_user}, votes: {c.votes}, confidence: {c.confidence:.2f}")
-
-
-@collaborative_group.command(name="recommend")
-@click.argument("entity")
-@click.option("--limit", type=int, default=10, help="Máximo de resultados")
-def collaborative_recommend(entity: str, limit: int) -> None:
-    """Obtiene conexiones recomendadas basadas en patrones colectivos.
-
-    Example: memo collaborative recommend MLX
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    recommendations = mem.collaborative.get_recommended_connections(entity, limit=limit)
-
-    console.print(f"[bold]Recommended connections for {entity}[/bold]")
-    for r in recommendations:
-        console.print(f"  {r.entity_a} --{r.relationship}--> {r.entity_b}")
-        console.print(f"    From: {r.from_user}, votes: {r.votes}, confidence: {r.confidence:.2f}")
-
-
-@collaborative_group.command(name="share-insight")
-@click.argument("user-id")
-@click.argument("content")
-def collaborative_share_insight(user_id: str, content: str) -> None:
-    """Comparte un insight con la comunidad.
-
-    Example: memo collaborative share-insight user123 "MLX is ideal for edge because..."
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    insight = mem.collaborative.share_insight(user_id, content)
-
-    console.print("[green]Insight shared[/green]")
-    console.print(f"Insight ID: {insight.insight_id}")
-    console.print(f"Content: {content[:100]}...")
-
-
-@collaborative_group.command(name="insights")
-@click.option("--limit", type=int, default=10, help="Máximo de resultados")
-def collaborative_insights(limit: int) -> None:
-    """Ver los insights más votados de la comunidad.
-
-    Example: memo collaborative insights
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    insights = mem.collaborative.get_top_insights(limit=limit)
-
-    console.print("[bold]Top insights[/bold]")
-    for i, insight in enumerate(insights, 1):
-        console.print(f"[cyan]{i}.[/cyan] {insight.content[:100]}...")
-        console.print(f"    Upvotes: {insight.upvotes}, Downvotes: {insight.downvotes}")
-
-
-# -- contradiction radar + dedupe ---------------------------------------------
-
-
-@cli.group(name="contradict")
-def contradict_group() -> None:
-    """Detect and triage contradictions / staleness across the corpus.
-
-    `scan` runs the LLM classifier over near-neighbor pairs and stores
-    contradictions in a sidecar DB. `list` shows open pairs. `triage`
-    walks them one by one and applies the user's verdict (fuse / keep
-    newer / dismiss / etc.).
-    """
-    pass
-
-
-def _short(text: str, n: int = 120) -> str:
-    text = (text or "").strip().replace("\n", " ")
-    return text if len(text) <= n else text[: n - 1] + "…"
-
-
-def _fmt_pair_header(rec_a, rec_b, pair) -> str:
-    rel = pair.relationship
-    color = "red" if rel == "contradiction" else "yellow"
-    return (
-        f"[bold {color}]{rel.upper()}[/bold {color}] "
-        f"conf={pair.confidence:.2f}  pair={pair.pair_id}"
-    )
-
-
-@contradict_group.command(name="scan")
-@click.option("--top-k", type=int, default=5,
-              help="Vec neighbors to consider per memoria (default: 5)")
-@click.option("--sim-floor", type=float, default=0.55,
-              help="Cosine floor; pairs below are skipped (default: 0.55)")
-@click.option("--confidence", type=float, default=0.7,
-              help="Min LLM confidence to store (default: 0.7)")
-@click.option("--min-days-apart", type=int, default=1,
-              help="Skip pairs whose updates are within N days (default: 1)")
-@click.option("--max-memorias", type=int, default=2000,
-              help="Cap on memorias visited (default: 2000)")
-@click.option("--max-pairs", type=int, default=500,
-              help="Cap on pairs sent to the LLM (default: 500)")
-@click.option("--since", help="Only scan memorias updated on/after this ISO date")
-@click.option("--type", "type_", help="Filter by memoria type")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def contradict_scan(
-    top_k: int, sim_floor: float, confidence: float, min_days_apart: int,
-    max_memorias: int, max_pairs: int, since: str | None, type_: str | None,
-    as_json: bool,
-) -> None:
-    """Scan the corpus for contradiction/evolution pairs.
-
-    Example: memo contradict scan --since 2026-04-01
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    console.print("[bold]Scanning corpus for contradictions…[/bold]")
-    last_idx = {"n": 0}
-
-    def progress(idx: int, total: int, title: str) -> None:
-        if idx == total or idx - last_idx["n"] >= 25:
-            console.print(f"[dim]  {idx}/{total}  {_short(title, 60)}[/dim]")
-            last_idx["n"] = idx
-
-    result = mem.contradict_scanner.scan_corpus(
-        top_k=top_k,
-        sim_floor=sim_floor,
-        confidence_threshold=confidence,
-        min_days_apart=min_days_apart,
-        max_memorias=max_memorias,
-        max_pairs=max_pairs,
-        since=since,
-        type_=type_,
-        progress=progress,
-    )
-
-    payload = {
-        "scanned_memorias": result.scanned_memorias,
-        "pairs_examined": result.pairs_examined,
-        "pairs_inserted": result.pairs_inserted,
-        "pairs_refreshed": result.pairs_refreshed,
-        "pairs_skipped_resolved": result.pairs_skipped_resolved,
-        "contradictions_found": result.contradictions_found,
-        "evolutions_found": result.evolutions_found,
-    }
-    if as_json:
-        click.echo(json.dumps(payload, indent=2))
-        return
-
-    table = Table(title="Scan summary")
-    table.add_column("metric")
-    table.add_column("value", justify="right")
-    for k, v in payload.items():
-        table.add_row(k.replace("_", " "), str(v))
-    console.print(table)
-
-    if result.contradictions_found or result.evolutions_found:
-        console.print(
-            "\n[green]→[/green] Run [cyan]memo contradict triage[/cyan] to resolve them."
-        )
-
-
-@contradict_group.command(name="list")
-@click.option("--limit", type=int, default=20, help="Max rows (default: 20)")
-@click.option("--min-confidence", type=float, default=0.0)
-@click.option("--relationship", type=click.Choice(["contradiction", "evolution"]),
-              help="Filter by relationship type")
-@click.option("--status", type=click.Choice(
-    ["open", "fused", "kept_newer", "kept_older", "evolved", "dismissed"]),
-    default="open", help="Filter by status (default: open)")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON")
-def contradict_list(
-    limit: int, min_confidence: float, relationship: str | None,
-    status: str, as_json: bool,
-) -> None:
-    """List contradiction pairs from the sidecar DB."""
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    if status == "open":
-        pairs = mem.contradict_store.list_open(
-            limit=limit, min_confidence=min_confidence, relationship=relationship,
-        )
-    else:
-        pairs = mem.contradict_store.list_all(status=status, limit=limit)
-        if relationship:
-            pairs = [p for p in pairs if p.relationship == relationship]
-
-    if as_json:
-        click.echo(json.dumps([p.__dict__ for p in pairs], indent=2, default=str))
-        return
-
-    if not pairs:
-        console.print(f"[green]No pairs in status '{status}'[/green]")
-        return
-
-    table = Table(title=f"Contradiction pairs · status={status}")
-    table.add_column("id", justify="right")
-    table.add_column("rel")
-    table.add_column("conf", justify="right")
-    table.add_column("a")
-    table.add_column("b")
-    table.add_column("rationale")
-    for p in pairs:
-        table.add_row(
-            str(p.pair_id),
-            p.relationship,
-            f"{p.confidence:.2f}",
-            p.memoria_id_a[:8],
-            p.memoria_id_b[:8],
-            _short(p.rationale, 70),
-        )
-    console.print(table)
-
-
-def _display_pair_excerpt(rec, label: str, *, stale_days: int = 180) -> None:
-    from memo.contradict import is_stale
-    age_marker = "  [red](stale)[/red]" if is_stale(rec.updated, stale_days) else ""
-    console.print(
-        f"[bold cyan]{label}[/bold cyan] · {rec.id[:8]} · "
-        f"[dim]{rec.type}[/dim] · updated={rec.updated[:10]}{age_marker}"
-    )
-    console.print(f"  [bold]{rec.title}[/bold]")
-    body = (rec.body or "").strip()
-    if len(body) > 600:
-        body = body[:599] + "…"
-    for line in body.splitlines():
-        console.print(f"  {line}")
-    console.print()
-
-
-_TRIAGE_HELP = """
-Actions for each pair:
-  f = fuse (LLM-merge both → new memoria, archive both)
-  n = newer wins (keep newer, delete older)
-  o = older wins (keep older, delete newer)
-  e = evolved (legitimate evolution, mark resolved, keep both)
-  d = dismiss (false positive)
-  s = skip (leave as open)
-  q = quit walker
-""".strip()
-
-
-@contradict_group.command(name="triage")
-@click.option("--limit", type=int, default=20,
-              help="Max pairs to walk in this session (default: 20)")
-@click.option("--min-confidence", type=float, default=0.7,
-              help="Skip pairs below this LLM confidence (default: 0.7)")
-@click.option("--relationship", type=click.Choice(["contradiction", "evolution"]),
-              help="Only walk pairs of this relationship type")
-@click.option("--stale-days", type=int, default=180,
-              help="Days threshold for the [stale] marker (default: 180)")
-@click.option("--yes-fuse", is_flag=True,
-              help="Auto-accept fuse without an extra confirmation prompt")
-def contradict_triage(
-    limit: int, min_confidence: float, relationship: str | None,
-    stale_days: int, yes_fuse: bool,
-) -> None:
-    """Interactive triage walker over open contradiction pairs.
-
-    Example: memo contradict triage --relationship contradiction
-    """
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-
-    pairs = mem.contradict_store.list_open(
-        limit=limit, min_confidence=min_confidence, relationship=relationship,
-    )
-    if not pairs:
-        console.print("[green]No open pairs to triage.[/green]")
-        return
-
-    console.print(f"[bold]Walking {len(pairs)} pair(s).[/bold] Type [cyan]?[/cyan] for help.")
-
-    for pair in pairs:
-        rec_a = mem.get(pair.memoria_id_a)
-        rec_b = mem.get(pair.memoria_id_b)
-        if rec_a is None or rec_b is None:
-            mem.contradict_store.resolve(
-                pair.pair_id, "dismissed",
-                note="auto: one side missing at triage time",
-            )
-            continue
-
-        # Orient newer as "B" so the walker is always presented with
-        # the same temporal layout (older on top, newer below).
-        if rec_a.updated > rec_b.updated:
-            rec_a, rec_b = rec_b, rec_a
-
-        console.print()
-        console.print(_fmt_pair_header(rec_a, rec_b, pair))
-        if pair.rationale:
-            console.print(f"[dim]rationale: {pair.rationale}[/dim]")
-        console.print()
-        _display_pair_excerpt(rec_a, "OLDER", stale_days=stale_days)
-        _display_pair_excerpt(rec_b, "NEWER", stale_days=stale_days)
-
-        while True:
-            choice = click.prompt(
-                "Action [f/n/o/e/d/s/q/?]",
-                type=str, default="s", show_default=False,
-            ).strip().lower()
-            if choice == "?":
-                console.print(_TRIAGE_HELP)
-                continue
-            if choice in {"f", "n", "o", "e", "d", "s", "q"}:
-                break
-            console.print("[red]Unknown action.[/red] Use ? for help.")
-
-        if choice == "q":
-            console.print("[dim]Stopping walker.[/dim]")
-            break
-        if choice == "s":
-            continue
-        if choice == "d":
-            note = click.prompt("note (optional)", default="", show_default=False) or None
-            mem.contradict_store.resolve(pair.pair_id, "dismissed", note=note)
-            console.print("[dim]dismissed.[/dim]")
-            continue
-        if choice == "e":
-            mem.contradict_store.resolve(pair.pair_id, "evolved")
-            console.print("[dim]marked as evolved.[/dim]")
-            continue
-        if choice == "n":
-            if click.confirm(f"Delete OLDER {rec_a.id[:8]}?", default=False):
-                mem.delete(rec_a.id)
-                mem.contradict_store.resolve(pair.pair_id, "kept_newer",
-                                             note=f"deleted older {rec_a.id}")
-                console.print(f"[green]kept newer.[/green] older {rec_a.id[:8]} deleted.")
-            continue
-        if choice == "o":
-            if click.confirm(f"Delete NEWER {rec_b.id[:8]}?", default=False):
-                mem.delete(rec_b.id)
-                mem.contradict_store.resolve(pair.pair_id, "kept_older",
-                                             note=f"deleted newer {rec_b.id}")
-                console.print(f"[green]kept older.[/green] newer {rec_b.id[:8]} deleted.")
-            continue
-        if choice == "f":
-            cluster = {
-                "cluster_id": pair.pair_id,
-                "relationship": "duplicate",
-                "rationale": pair.rationale,
-                "members": [
-                    {
-                        "id": rec_a.id, "title": rec_a.title,
-                        "updated": rec_a.updated,
-                        "body_preview": (rec_a.body or "")[:400],
-                    },
-                    {
-                        "id": rec_b.id, "title": rec_b.title,
-                        "updated": rec_b.updated,
-                        "body_preview": (rec_b.body or "")[:400],
-                    },
-                ],
-            }
-            proposal = mem.consolidator.propose_merge(cluster)
-            if proposal is None:
-                console.print("[red]LLM declined to propose a merge. Skipping.[/red]")
-                continue
-            console.print(
-                f"[bold]Proposed merged title:[/bold] {proposal.merged_title}"
-            )
-            console.print(f"[dim]strategy={proposal.merge_strategy}[/dim]")
-            console.print(f"[dim]rationale={proposal.rationale}[/dim]")
-            if yes_fuse or click.confirm("Apply this merge?", default=False):
-                merge_result = mem.consolidator.apply_merge(proposal, dry_run=False)
-                mem.contradict_store.resolve(
-                    pair.pair_id, "fused",
-                    note=f"merged into {merge_result.merged_id}",
-                )
-                console.print(
-                    f"[green]fused →[/green] {merge_result.merged_id[:8] if merge_result.merged_id else 'n/a'}"
-                )
-            continue
-
-    stats = mem.contradict_store.stats()
-    console.print()
-    console.print(f"[bold]Session stats:[/bold] {stats}")
-
-
-@contradict_group.command(name="stats")
-def contradict_stats() -> None:
-    """Show counts of pairs grouped by status."""
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-    stats = mem.contradict_store.stats()
-    if not stats:
-        console.print("[dim]No pairs recorded yet. Run `memo contradict scan` first.[/dim]")
-        return
-    table = Table(title="Contradiction pairs by status")
-    table.add_column("status")
-    table.add_column("count", justify="right")
-    for k, v in sorted(stats.items()):
-        table.add_row(k, str(v))
-    console.print(table)
-
-
-@contradict_group.command(name="reopen")
-@click.argument("pair_id", type=int)
-def contradict_reopen(pair_id: int) -> None:
-    """Send a resolved pair back to the open queue."""
-    cfg = Config.from_env()
-    mem = _get_memory(cfg)
-    if mem.contradict_store.reopen(pair_id):
-        console.print(f"[green]pair {pair_id} reopened.[/green]")
-    else:
-        console.print(f"[red]pair {pair_id} not found.[/red]")
 
 
 # -- duplicate detection (exact-ish near-dups, no LLM gate) -------------------
